Handle typed expiry in recall
