+++ internal/service/recallservice/recall.go
@@ -28,7 +28,6 @@ import (
 	"errors"
 	"fmt"
 	"sort"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -597,6 +596,7 @@ func (s *recallService) enrichTierHits(ctx context.Context, profileID string, li
 				hydrated = append(hydrated, hydratedFactRecallCandidate{Fact: f, AuthorityState: authorityState})
 			}
 			evidenceFragments := s.batchHydrateEvidenceFragments(ctx, profileID, factEvidenceSets(hydrated))
+			temporalFrame := typedCurrentnessTemporalFrameForFacts(req.Query, hydrated, evidenceFragments)
 			for _, candidate := range hydrated {
 				f := candidate.Fact
 				f.AuthorityState = candidate.AuthorityState
@@ -604,7 +604,7 @@ func (s *recallService) enrichTierHits(ctx context.Context, profileID string, li
 				if candidate.AuthorityState != "authoritative" {
 					tier = TierConflict
 				}
-				temporalRank := temporalRankTimeForRecallWithEvidence(req.Query, f.ValidFrom, f.RecordedAt, f.Evidence, evidenceFragments, f.Subject, f.Predicate, f.Object)
+				temporalRank := typedTemporalRankTimeForRecallWithEvidence(req.Query, f.ValidFrom, f.ValidTo, f.RecordedAt, f.Evidence, evidenceFragments, temporalFrame, f.Subject, f.Predicate, f.Object)
 				if !req.IncludeEvidence {
 					factCopy := *f
 					factCopy.Evidence = nil
@@ -666,8 +666,9 @@ func (s *recallService) enrichTierHits(ctx context.Context, profileID string, li
 				hydrated = append(hydrated, c)
 			}
 			evidenceFragments := s.batchHydrateEvidenceFragments(ctx, profileID, claimEvidenceSets(hydrated))
+			temporalFrame := typedCurrentnessTemporalFrameForClaims(req.Query, hydrated, evidenceFragments)
 			for _, c := range hydrated {
-				temporalRank := temporalRankTimeForRecallWithEvidence(req.Query, c.ValidFrom, c.RecordedAt, c.Evidence, evidenceFragments, c.Subject, c.Predicate, c.Object)
+				temporalRank := typedTemporalRankTimeForRecallWithEvidence(req.Query, c.ValidFrom, c.ValidTo, c.RecordedAt, c.Evidence, evidenceFragments, temporalFrame, c.Subject, c.Predicate, c.Object)
 				if !req.IncludeEvidence {
 					claimCopy := *c
 					claimCopy.Evidence = nil
@@ -880,1279 +881,6 @@ func mergeRRFEntryTimes(entry *rrfEntry, createdAt, updatedAt time.Time) {
 	if !updatedAt.IsZero() && (entry.UpdatedAt.IsZero() || updatedAt.After(entry.UpdatedAt)) {
 		entry.UpdatedAt = updatedAt
 	}
-}
-
-func applyIdentifierSpecificityAdjustments(query string, entries []rrfEntry) {
-	queryText := rerankText(query)
-	if !isUnitValueQueryText(queryText) {
-		return
-	}
-	if len(rerankIdentifiers(queryText)) == 0 {
-		return
-	}
-	for i := range entries {
-		entries[i].FinalScore += identifierSpecificityAdjustment(queryText, entries[i].Content)
-	}
-}
-
-func isUnitValueQueryText(queryText string) bool {
-	return strings.Contains(queryText, " timeout ") &&
-		strings.Contains(queryText, " job ") &&
-		(strings.Contains(queryText, " use ") || strings.Contains(queryText, " should "))
-}
-
-func identifierSpecificityAdjustment(queryText, content string) float64 {
-	contentText := rerankText(content)
-	if queryText == "" || contentText == "" {
-		return 0
-	}
-	if !matchesQueryIdentifiers(queryText, contentText) {
-		return 0
-	}
-	return 0.004
-}
-
-func applyCurrentnessAdjustments(query string, entries []rrfEntry) {
-	if !isCurrentnessQuery(query) {
-		return
-	}
-	temporalFrame := currentnessTemporalFrameFor(query, entries)
-	for i := range entries {
-		lexicalAdjustment := currentnessAdjustment(query, entries[i].Content)
-		if temporalFrame.hasContentDate && lexicalAdjustment > 0 && latestCurrentnessTemporalDateInEntry(entries[i]).IsZero() {
-			lexicalAdjustment = 0
-		}
-		entries[i].FinalScore += lexicalAdjustment
-		entries[i].FinalScore += currentnessTemporalAdjustment(query, entries[i], temporalFrame)
-		entries[i].FinalScore += expiredValidityAdjustment(query, entries[i], temporalFrame)
-		if entries[i].FinalScore < 0 {
-			entries[i].FinalScore = 0
-		}
-	}
-}
-
-func applyCueAdjustments(query string, entries []rrfEntry) {
-	if !isSelectionRecallQuery(query) {
-		return
-	}
-	frame := selectionCueFrameFor(query, entries)
-	for i := range entries {
-		entries[i].FinalScore += cueAdjustment(query, entries[i].Content)
-		entries[i].FinalScore += historicalSelectionAdjustment(query, entries[i].Content, frame)
-		if entries[i].FinalScore < 0 {
-			entries[i].FinalScore = 0
-		}
-	}
-}
-
-func applyAuthorityAdjustments(query string, entries []rrfEntry) {
-	if !isAuthorityRecallQuery(query) {
-		return
-	}
-	for i := range entries {
-		entries[i].FinalScore += authorityAdjustment(query, entries[i].Content)
-		if entries[i].FinalScore < 0 {
-			entries[i].FinalScore = 0
-		}
-	}
-}
-
-func currentnessAdjustment(query, content string) float64 {
-	queryText := rerankText(query)
-	contentText := rerankText(content)
-	if queryText == "" || contentText == "" {
-		return 0
-	}
-
-	adjustment := 0.0
-	hasCurrentCue := containsAnyRerankCue(contentText, currentnessPositiveCues)
-	if hasCurrentCue && rerankMatchesQueryIdentifiers(queryText, contentText) {
-		adjustment += 0.024
-	}
-	if containsAnyRerankCue(contentText, currentnessStrongStaleCues) {
-		if hasCurrentCue {
-			adjustment -= 0.006
-		} else {
-			adjustment -= 0.024
-		}
-	}
-	if containsAnyRerankCue(contentText, currentnessWeakStaleCues) {
-		adjustment -= 0.012
-	}
-
-	if adjustment > 0.028 {
-		return 0.028
-	}
-	if adjustment < -0.028 {
-		return -0.028
-	}
-	return adjustment
-}
-
-type currentnessTemporalFrame struct {
-	hasContentDate       bool
-	newestContentDate    time.Time
-	useFragmentTimestamp bool
-	newestFragmentTime   time.Time
-}
-
-func currentnessTemporalAdjustment(query string, entry rrfEntry, frame currentnessTemporalFrame) float64 {
-	if !frame.hasContentDate && !frame.useFragmentTimestamp {
-		return 0
-	}
-	queryText := rerankText(query)
-	contentText := rerankText(entry.Content)
-	if queryText == "" || contentText == "" || !rerankMatchesQueryIdentifiers(queryText, contentText) {
-		return 0
-	}
-
-	if frame.hasContentDate {
-		contentDate := latestCurrentnessTemporalDateInEntry(entry)
-		if contentDate.IsZero() {
-			return -0.006
-		}
-		return contentDateTemporalDelta(frame.newestContentDate.Sub(contentDate))
-	}
-
-	fragmentTime := latestFragmentTimestamp(entry.CreatedAt, entry.UpdatedAt)
-	if fragmentTime.IsZero() {
-		return 0
-	}
-	return fragmentTimestampTemporalDelta(frame.newestFragmentTime.Sub(fragmentTime))
-}
-
-func contentDateTemporalDelta(age time.Duration) float64 {
-	if age < 0 {
-		age = 0
-	}
-	switch {
-	case age == 0:
-		return 0.028
-	case age <= 72*time.Hour:
-		return 0.014
-	case age >= 7*24*time.Hour:
-		return -0.026
-	case age >= 24*time.Hour:
-		return -0.018
-	default:
-		return 0
-	}
-}
-
-func fragmentTimestampTemporalDelta(age time.Duration) float64 {
-	if age < 0 {
-		age = 0
-	}
-	switch {
-	case age == 0:
-		return 0.014
-	case age <= 72*time.Hour:
-		return 0.007
-	case age >= 7*24*time.Hour:
-		return -0.014
-	case age >= 24*time.Hour:
-		return -0.010
-	default:
-		return 0
-	}
-}
-
-func expiredValidityAdjustment(query string, entry rrfEntry, frame currentnessTemporalFrame) float64 {
-	queryText := rerankText(query)
-	contentText := rerankText(entry.Content)
-	if queryText == "" || contentText == "" || !rerankMatchesQueryIdentifiers(queryText, contentText) {
-		return 0
-	}
-	validityEnd := latestValidityEndDateInEntry(entry)
-	if validityEnd.IsZero() {
-		return 0
-	}
-	asOf := currentnessAsOfTime(query, entry, frame)
-	if asOf.IsZero() || !utcDate(asOf).After(utcDate(validityEnd)) {
-		return 0
-	}
-	return -0.04
-}
-
-func currentnessAsOfTime(query string, entry rrfEntry, frame currentnessTemporalFrame) time.Time {
-	if queryDate := latestTemporalDateInText(query, latestFragmentTimestamp(entry.CreatedAt, entry.UpdatedAt)); !queryDate.IsZero() {
-		return queryDate
-	}
-	if !frame.newestContentDate.IsZero() {
-		return frame.newestContentDate
-	}
-	if !frame.newestFragmentTime.IsZero() {
-		return frame.newestFragmentTime
-	}
-	return latestFragmentTimestamp(entry.CreatedAt, entry.UpdatedAt)
-}
-
-func currentnessTemporalFrameFor(query string, entries []rrfEntry) currentnessTemporalFrame {
-	queryText := rerankText(query)
-	var frame currentnessTemporalFrame
-	var oldestFragmentTime time.Time
-	for _, entry := range entries {
-		contentText := rerankText(entry.Content)
-		if queryText == "" || contentText == "" || !rerankMatchesQueryIdentifiers(queryText, contentText) {
-			continue
-		}
-		if contentDate := latestCurrentnessTemporalDateInEntry(entry); !contentDate.IsZero() {
-			frame.hasContentDate = true
-			if frame.newestContentDate.IsZero() || contentDate.After(frame.newestContentDate) {
-				frame.newestContentDate = contentDate
-			}
-		}
-		fragmentTime := latestFragmentTimestamp(entry.CreatedAt, entry.UpdatedAt)
-		if fragmentTime.IsZero() {
-			continue
-		}
-		if oldestFragmentTime.IsZero() || fragmentTime.Before(oldestFragmentTime) {
-			oldestFragmentTime = fragmentTime
-		}
-		if frame.newestFragmentTime.IsZero() || fragmentTime.After(frame.newestFragmentTime) {
-			frame.newestFragmentTime = fragmentTime
-		}
-	}
-	if !frame.hasContentDate && !oldestFragmentTime.IsZero() && frame.newestFragmentTime.Sub(oldestFragmentTime) >= 24*time.Hour {
-		frame.useFragmentTimestamp = true
-	}
-	return frame
-}
-
-func latestFragmentTimestamp(createdAt, updatedAt time.Time) time.Time {
-	latest := createdAt
-	if updatedAt.After(latest) {
-		latest = updatedAt
-	}
-	if latest.IsZero() {
-		return time.Time{}
-	}
-	return latest.UTC()
-}
-
-func temporalRankTimeForRecall(query string, validFrom *time.Time, recordedAt time.Time, contentParts ...string) time.Time {
-	return temporalRankTimeForRecallWithEvidence(query, validFrom, recordedAt, nil, nil, contentParts...)
-}
-
-func temporalRankTimeForRecallWithEvidence(query string, validFrom *time.Time, recordedAt time.Time, evidence []domain.Evidence, evidenceFragments map[string]*domain.Fragment, contentParts ...string) time.Time {
-	if !isCurrentnessQuery(query) {
-		return time.Time{}
-	}
-	if validFrom != nil && !validFrom.IsZero() {
-		return validFrom.UTC()
-	}
-	if contentDate := latestTemporalDateInText(strings.Join(contentParts, " "), recordedAt); !contentDate.IsZero() {
-		return contentDate
-	}
-	if evidenceDate := latestTemporalDateInEvidence(query, evidence, evidenceFragments); !evidenceDate.IsZero() {
-		return evidenceDate
-	}
-	if !recordedAt.IsZero() {
-		return recordedAt.UTC()
-	}
-	return time.Time{}
-}
-
-func factMatchesQueryIdentifiers(query string, fact *domain.Fact) bool {
-	if fact == nil {
-		return false
-	}
-	return knowledgeTripleMatchesQueryIdentifiers(query, fact.Subject, fact.Predicate, fact.Object)
-}
-
-func claimMatchesQueryIdentifiers(query string, claim *domain.Claim) bool {
-	if claim == nil {
-		return false
-	}
-	return knowledgeTripleMatchesQueryIdentifiers(query, claim.Subject, claim.Predicate, claim.Object)
-}
-
-func knowledgeTripleMatchesQueryIdentifiers(query string, parts ...string) bool {
-	queryText := rerankText(query)
-	if len(rerankIdentifiers(queryText)) == 0 {
-		return true
-	}
-	return rerankMatchesQueryIdentifiers(queryText, rerankText(strings.Join(parts, " ")))
-}
-func latestISODateInText(value string) time.Time {
-	var latest time.Time
-	for _, field := range strings.Fields(value) {
-		token := strings.Trim(field, ".,:;!?()[]{}\"'")
-		if len(token) != len("2006-01-02") {
-			continue
-		}
-		parsed, err := time.Parse("2006-01-02", token)
-		if err != nil {
-			continue
-		}
-		parsed = parsed.UTC()
-		if latest.IsZero() || parsed.After(latest) {
-			latest = parsed
-		}
-	}
-	return latest
-}
-
-func latestNumericDateInText(value string, anchor time.Time) time.Time {
-	anchorYear := 0
-	if !anchor.IsZero() {
-		anchorYear = anchor.UTC().Year()
-	}
-	latest := time.Time{}
-	for _, field := range strings.Fields(value) {
-		token := strings.Trim(strings.ToLower(field), ".,:;!?()[]{}\"'")
-		candidate, ok := numericDateToken(token, anchorYear)
-		if !ok {
-			continue
-		}
-		if latest.IsZero() || candidate.After(latest) {
-			latest = candidate
-		}
-	}
-	return latest
-}
-
-func numericDateToken(token string, anchorYear int) (time.Time, bool) {
-	parts := strings.Split(token, "/")
-	if len(parts) != 2 && len(parts) != 3 {
-		return time.Time{}, false
-	}
-	nums := make([]int, len(parts))
-	for i, part := range parts {
-		if part == "" {
-			return time.Time{}, false
-		}
-		value, err := strconv.Atoi(part)
-		if err != nil {
-			return time.Time{}, false
-		}
-		nums[i] = value
-	}
-	if len(parts) == 3 {
-		if year, ok := parseYear(parts[0]); ok {
-			return calendarDate(year, time.Month(nums[1]), nums[2])
-		}
-		year, ok := parseYear(parts[2])
-		if !ok {
-			return time.Time{}, false
-		}
-		month, day, ok := unambiguousNumericMonthDay(nums[0], nums[1])
-		if !ok {
-			return time.Time{}, false
-		}
-		return calendarDate(year, time.Month(month), day)
-	}
-	if anchorYear == 0 {
-		return time.Time{}, false
-	}
-	month, day, ok := unambiguousNumericMonthDay(nums[0], nums[1])
-	if !ok {
-		return time.Time{}, false
-	}
-	return calendarDate(anchorYear, time.Month(month), day)
-}
-
-func unambiguousNumericMonthDay(first, second int) (int, int, bool) {
-	switch {
-	case first >= 1 && first <= 12 && second > 12 && second <= 31:
-		return first, second, true
-	case first > 12 && first <= 31 && second >= 1 && second <= 12:
-		return second, first, true
-	default:
-		return 0, 0, false
-	}
-}
-
-func latestTemporalDateInEntry(entry rrfEntry) time.Time {
-	return latestTemporalDateInText(entry.Content, latestFragmentTimestamp(entry.CreatedAt, entry.UpdatedAt))
-}
-
-func latestCurrentnessTemporalDateInEntry(entry rrfEntry) time.Time {
-	latest := latestTemporalDateInEntry(entry)
-	if latest.IsZero() {
-		return time.Time{}
-	}
-	validityEnd := latestValidityEndDateInEntry(entry)
-	if !validityEnd.IsZero() && utcDate(latest).Equal(utcDate(validityEnd)) {
-		return time.Time{}
-	}
-	return latest
-}
-
-func latestValidityEndDateInEntry(entry rrfEntry) time.Time {
-	return latestValidityEndDateInText(entry.Content, latestFragmentTimestamp(entry.CreatedAt, entry.UpdatedAt))
-}
-
-func latestTemporalDateInEvidence(query string, evidence []domain.Evidence, evidenceFragments map[string]*domain.Fragment) time.Time {
-	if len(evidence) == 0 || len(evidenceFragments) == 0 {
-		return time.Time{}
-	}
-	queryText := rerankText(query)
-	latest := time.Time{}
-	for _, item := range evidence {
-		if item.FragmentID == "" {
-			continue
-		}
-		fragment := evidenceFragments[item.FragmentID]
-		if fragment == nil {
-			continue
-		}
-		contentText := rerankText(fragment.Content)
-		if queryText == "" || contentText == "" || !rerankMatchesQueryIdentifiers(queryText, contentText) {
-			continue
-		}
-		candidate := latestTemporalDateInEntry(rrfEntry{
-			Content:   fragment.Content,
-			CreatedAt: fragment.CreatedAt,
-			UpdatedAt: fragment.UpdatedAt,
-		})
-		if !candidate.IsZero() && (latest.IsZero() || candidate.After(latest)) {
-			latest = candidate
-		}
-	}
-	return latest
-}
-
-func latestTemporalDateInText(value string, anchor time.Time) time.Time {
-	latest := latestISODateInText(value)
-	numeric := latestNumericDateInText(value, anchor)
-	if !numeric.IsZero() && (latest.IsZero() || numeric.After(latest)) {
-		latest = numeric
-	}
-	monthName := latestMonthNameDateInText(value, anchor)
-	if !monthName.IsZero() && (latest.IsZero() || monthName.After(latest)) {
-		latest = monthName
-	}
-	relative := latestRelativeDateInText(value, anchor)
-	if !relative.IsZero() && (latest.IsZero() || relative.After(latest)) {
-		latest = relative
-	}
-	weekday := latestWeekdayDateInText(value, anchor)
-	if !weekday.IsZero() && (latest.IsZero() || weekday.After(latest)) {
-		latest = weekday
-	}
-	return latest
-}
-
-func latestValidityEndDateInText(value string, anchor time.Time) time.Time {
-	text := rerankText(value)
-	if text == "" {
-		return time.Time{}
-	}
-	fields := strings.Fields(text)
-	latest := time.Time{}
-	for i := range fields {
-		if !validityEndCueAt(fields, i) {
-			continue
-		}
-		for j := i + 1; j < len(fields) && j <= i+4; j++ {
-			candidate, ok := temporalDateAtFields(fields, j, anchor)
-			if !ok {
-				continue
-			}
-			if latest.IsZero() || candidate.After(latest) {
-				latest = candidate
-			}
-			break
-		}
-	}
-	return latest
-}
-
-func validityEndCueAt(fields []string, index int) bool {
-	if index < 0 || index >= len(fields) {
-		return false
-	}
-	field := fields[index]
-	if field == "valid" && index+1 < len(fields) && validityEndPreposition(fields[index+1]) {
-		return true
-	}
-	if validityEndPreposition(field) && index > 0 && fields[index-1] == "valid" {
-		return true
-	}
-	switch field {
-	case "expires", "expire", "expired", "expiration", "ends", "ended", "ending", "sunsets", "sunset", "retired":
-		return true
-	default:
-		return false
-	}
-}
-
-func validityEndPreposition(token string) bool {
-	switch token {
-	case "through", "until", "to", "before":
-		return true
-	default:
-		return false
-	}
-}
-
-func temporalDateAtFields(fields []string, index int, anchor time.Time) (time.Time, bool) {
-	if index < 0 || index >= len(fields) {
-		return time.Time{}, false
-	}
-	if parsed, err := time.Parse("2006-01-02", fields[index]); err == nil {
-		return parsed.UTC(), true
-	}
-	anchorYear := 0
-	if !anchor.IsZero() {
-		anchorYear = anchor.UTC().Year()
-	}
-	if parsed, ok := numericDateToken(fields[index], anchorYear); ok {
-		return parsed, true
-	}
-	if month, ok := monthNameNumber(fields[index]); ok {
-		if index+1 < len(fields) {
-			if day, ok := parseDayOfMonth(fields[index+1]); ok {
-				year := anchorYear
-				if index+2 < len(fields) {
-					if parsedYear, ok := parseYear(fields[index+2]); ok {
-						year = parsedYear
-					}
-				}
-				return calendarDate(year, month, day)
-			}
-		}
-	}
-	if day, ok := parseDayOfMonth(fields[index]); ok && index+1 < len(fields) {
-		if month, ok := monthNameNumber(fields[index+1]); ok {
-			year := anchorYear
-			if index+2 < len(fields) {
-				if parsedYear, ok := parseYear(fields[index+2]); ok {
-					year = parsedYear
-				}
-			}
-			return calendarDate(year, month, day)
-		}
-	}
-	return time.Time{}, false
-}
-
-func latestMonthNameDateInText(value string, anchor time.Time) time.Time {
-	text := rerankText(value)
-	if text == "" {
-		return time.Time{}
-	}
-	anchorYear := 0
-	if !anchor.IsZero() {
-		anchorYear = anchor.UTC().Year()
-	}
-	latest := time.Time{}
-	fields := strings.Fields(text)
-	for i, field := range fields {
-		month, ok := monthNameNumber(field)
-		if !ok {
-			continue
-		}
-		if i+1 < len(fields) {
-			if day, ok := parseDayOfMonth(fields[i+1]); ok {
-				year := anchorYear
-				if i+2 < len(fields) {
-					if parsedYear, ok := parseYear(fields[i+2]); ok {
-						year = parsedYear
-					}
-				}
-				if candidate, ok := calendarDate(year, month, day); ok && (latest.IsZero() || candidate.After(latest)) {
-					latest = candidate
-				}
-			}
-		}
-		if i > 0 {
-			if day, ok := parseDayOfMonth(fields[i-1]); ok {
-				year := anchorYear
-				if i+1 < len(fields) {
-					if parsedYear, ok := parseYear(fields[i+1]); ok {
-						year = parsedYear
-					}
-				}
-				if candidate, ok := calendarDate(year, month, day); ok && (latest.IsZero() || candidate.After(latest)) {
-					latest = candidate
-				}
-			}
-		}
-	}
-	return latest
-}
-
-func latestRelativeDateInText(value string, anchor time.Time) time.Time {
-	if anchor.IsZero() {
-		return time.Time{}
-	}
-	text := rerankText(value)
-	if text == "" {
-		return time.Time{}
-	}
-	anchorDate := utcDate(anchor)
-	latest := time.Time{}
-	add := func(candidate time.Time) {
-		if !candidate.IsZero() && (latest.IsZero() || candidate.After(latest)) {
-			latest = candidate
-		}
-	}
-	fields := strings.Fields(text)
-	for i, field := range fields {
-		switch field {
-		case "today":
-			add(anchorDate)
-		case "yesterday":
-			add(anchorDate.AddDate(0, 0, -1))
-		case "ago":
-			if i >= 2 && (fields[i-1] == "day" || fields[i-1] == "days") {
-				if days, ok := relativeDayCount(fields[i-2]); ok {
-					add(anchorDate.AddDate(0, 0, -days))
-				}
-			}
-		case "week":
-			if i >= 1 && fields[i-1] == "last" {
-				add(anchorDate.AddDate(0, 0, -7))
-			}
-		}
-	}
-	return latest
-}
-
-func latestWeekdayDateInText(value string, anchor time.Time) time.Time {
-	if anchor.IsZero() {
-		return time.Time{}
-	}
-	text := rerankText(value)
-	if text == "" {
-		return time.Time{}
-	}
-	anchorDate := utcDate(anchor)
-	latest := time.Time{}
-	fields := strings.Fields(text)
-	for i, field := range fields {
-		weekday, ok := weekdayNumber(field)
-		if !ok {
-			continue
-		}
-		if i > 0 && fields[i-1] == "next" {
-			continue
-		}
-		daysBack := (int(anchorDate.Weekday()) - int(weekday) + 7) % 7
-		if i > 0 && fields[i-1] == "last" && daysBack == 0 {
-			daysBack = 7
-		}
-		candidate := anchorDate.AddDate(0, 0, -daysBack)
-		if latest.IsZero() || candidate.After(latest) {
-			latest = candidate
-		}
-	}
-	return latest
-}
-
-func monthNameNumber(token string) (time.Month, bool) {
-	switch token {
-	case "jan", "january":
-		return time.January, true
-	case "feb", "february":
-		return time.February, true
-	case "mar", "march":
-		return time.March, true
-	case "apr", "april":
-		return time.April, true
-	case "may":
-		return time.May, true
-	case "jun", "june":
-		return time.June, true
-	case "jul", "july":
-		return time.July, true
-	case "aug", "august":
-		return time.August, true
-	case "sep", "sept", "september":
-		return time.September, true
-	case "oct", "october":
-		return time.October, true
-	case "nov", "november":
-		return time.November, true
-	case "dec", "december":
-		return time.December, true
-	default:
-		return 0, false
-	}
-}
-
-func weekdayNumber(token string) (time.Weekday, bool) {
-	switch token {
-	case "sun", "sunday":
-		return time.Sunday, true
-	case "mon", "monday":
-		return time.Monday, true
-	case "tue", "tues", "tuesday":
-		return time.Tuesday, true
-	case "wed", "wednesday":
-		return time.Wednesday, true
-	case "thu", "thur", "thurs", "thursday":
-		return time.Thursday, true
-	case "fri", "friday":
-		return time.Friday, true
-	case "sat", "saturday":
-		return time.Saturday, true
-	default:
-		return 0, false
-	}
-}
-
-func parseDayOfMonth(token string) (int, bool) {
-	day, err := strconv.Atoi(strings.TrimSuffix(strings.TrimSuffix(strings.TrimSuffix(strings.TrimSuffix(token, "st"), "nd"), "rd"), "th"))
-	if err != nil || day < 1 || day > 31 {
-		return 0, false
-	}
-	return day, true
-}
-
-func parseYear(token string) (int, bool) {
-	year, err := strconv.Atoi(token)
-	if err != nil || year < 1900 || year > 3000 {
-		return 0, false
-	}
-	return year, true
-}
-
-func calendarDate(year int, month time.Month, day int) (time.Time, bool) {
-	if year == 0 || month < time.January || month > time.December || day < 1 || day > daysInMonth(year, month) {
-		return time.Time{}, false
-	}
-	return time.Date(year, month, day, 0, 0, 0, 0, time.UTC), true
-}
-
-func daysInMonth(year int, month time.Month) int {
-	return time.Date(year, month+1, 0, 0, 0, 0, 0, time.UTC).Day()
-}
-
-func utcDate(value time.Time) time.Time {
-	value = value.UTC()
-	year, month, day := value.Date()
-	return time.Date(year, month, day, 0, 0, 0, 0, time.UTC)
-}
-
-func relativeDayCount(token string) (int, bool) {
-	switch token {
-	case "one", "a", "1":
-		return 1, true
-	case "two", "2":
-		return 2, true
-	case "three", "3":
-		return 3, true
-	case "four", "4":
-		return 4, true
-	case "five", "5":
-		return 5, true
-	case "six", "6":
-		return 6, true
-	case "seven", "7":
-		return 7, true
-	default:
-		return 0, false
-	}
-}
-
-func cueAdjustment(query, content string) float64 {
-	queryText := rerankText(query)
-	contentText := rerankText(content)
-	if queryText == "" || contentText == "" {
-		return 0
-	}
-
-	adjustment := 0.0
-	boostAllowed := matchesQueryIdentifiers(queryText, contentText)
-	if boostAllowed {
-		if containsAnyCue(contentText, directiveCues) {
-			adjustment += 0.022
-		}
-		if containsAnyCue(contentText, canonicalCues) {
-			adjustment += 0.004
-		}
-	}
-	if containsAnyCue(contentText, strongDisqualifierCues) {
-		adjustment -= 0.022
-	}
-	if containsAnyCue(contentText, weakDisqualifierCues) {
-		adjustment -= 0.012
-	}
-	if boostAllowed && containsAnyCue(queryText, conditionalQueryCues) && containsAnyCue(contentText, conditionalQueryCues) && containsAnyCue(contentText, directiveCues) {
-		adjustment += 0.006
-	}
-
-	if adjustment > 0.026 {
-		return 0.026
-	}
-	if adjustment < -0.026 {
-		return -0.026
-	}
-	return adjustment
-}
-
-type selectionCueFrame struct {
-	hasDirectiveMatch bool
-}
-
-func selectionCueFrameFor(query string, entries []rrfEntry) selectionCueFrame {
-	queryText := rerankText(query)
-	var frame selectionCueFrame
-	for _, entry := range entries {
-		contentText := rerankText(entry.Content)
-		if queryText == "" || contentText == "" || !matchesQueryIdentifiers(queryText, contentText) {
-			continue
-		}
-		if containsAnyCue(contentText, directiveCues) {
-			frame.hasDirectiveMatch = true
-			return frame
-		}
-	}
-	return frame
-}
-
-func historicalSelectionAdjustment(query, content string, frame selectionCueFrame) float64 {
-	if !frame.hasDirectiveMatch {
-		return 0
-	}
-	queryText := rerankText(query)
-	contentText := rerankText(content)
-	if queryText == "" || contentText == "" || !matchesQueryIdentifiers(queryText, contentText) {
-		return 0
-	}
-	if historicalSelectionActionCue(contentText) {
-		return -0.034
-	}
-	return 0
-}
-
-func historicalSelectionActionCue(contentText string) bool {
-	return strings.Contains(contentText, " before ") && strings.Contains(contentText, " used ")
-}
-
-func authorityAdjustment(query, content string) float64 {
-	queryText := rerankText(query)
-	contentText := rerankText(content)
-	if queryText == "" || contentText == "" {
-		return 0
-	}
-
-	adjustment := 0.0
-	boostAllowed := authorityMatchesQueryIdentifiers(queryText, contentText)
-	if boostAllowed {
-		if containsAnyAuthorityCue(contentText, authorityPositiveCues) {
-			adjustment += 0.026
-		}
-		if containsAnyAuthorityCue(contentText, authorityDirectiveCues) {
-			adjustment += 0.006
-		}
-	}
-	if containsAnyAuthorityCue(contentText, authorityStrongNegativeCues) {
-		adjustment -= 0.028
-	}
-	if containsAnyAuthorityCue(contentText, authorityWeakNegativeCues) {
-		adjustment -= 0.018
-	}
-
-	if adjustment > 0.034 {
-		return 0.034
-	}
-	if adjustment < -0.034 {
-		return -0.034
-	}
-	return adjustment
-}
-
-func isCurrentnessQuery(query string) bool {
-	text := rerankText(query)
-	return strings.Contains(text, " current ") ||
-		strings.Contains(text, " as of ") ||
-		strings.Contains(text, " now ") ||
-		strings.Contains(text, " latest ") ||
-		strings.Contains(text, " active ")
-}
-
-func isSelectionRecallQuery(query string) bool {
-	text := rerankText(query)
-	return strings.Contains(text, " which ") && (strings.Contains(text, " use ") || strings.Contains(text, " should "))
-}
-
-func rerankMatchesQueryIdentifiers(queryText, contentText string) bool {
-	identifiers := rerankIdentifiers(queryText)
-	return identifiersMatchContent(identifiers, contentText)
-}
-
-func matchesQueryIdentifiers(queryText, contentText string) bool {
-	return rerankMatchesQueryIdentifiers(queryText, contentText)
-}
-
-func isAuthorityRecallQuery(query string) bool {
-	text := rerankText(query)
-	return strings.Contains(text, " authoritative ") ||
-		strings.Contains(text, " canonical ") ||
-		strings.Contains(text, " require ") ||
-		strings.Contains(text, " requires ") ||
-		strings.Contains(text, " required ")
-}
-
-func isEvidenceSourceQuery(query string) bool {
-	text := rerankText(query)
-	return strings.Contains(text, " which source ") ||
-		strings.Contains(text, " what source ") ||
-		strings.Contains(text, " source note ") ||
-		strings.Contains(text, " source document ") ||
-		strings.Contains(text, " supporting evidence ") ||
-		strings.Contains(text, " raw evidence ") ||
-		strings.Contains(text, " original evidence ") ||
-		strings.Contains(text, " raw fragment ") ||
-		strings.Contains(text, " source fragment ") ||
-		strings.Contains(text, " which note ") ||
-		strings.Contains(text, " what note ") ||
-		strings.Contains(text, " note says ") ||
-		strings.Contains(text, " note said ") ||
-		strings.Contains(text, " mentioned ")
-}
-
-func authorityMatchesQueryIdentifiers(queryText, contentText string) bool {
-	identifiers := rerankIdentifiers(queryText)
-	return identifiersMatchContent(identifiers, contentText)
-}
-
-func identifiersMatchContent(identifiers []string, contentText string) bool {
-	if len(identifiers) == 0 {
-		return true
-	}
-	for _, identifier := range identifiers {
-		if !strings.Contains(contentText, " "+identifier+" ") {
-			return false
-		}
-	}
-	return true
-}
-
-func rerankIdentifiers(text string) []string {
-	fields := strings.Fields(text)
-	out := make([]string, 0, len(fields))
-	seen := make(map[string]struct{}, len(fields))
-	for _, field := range fields {
-		if _, ok := seen[field]; ok {
-			continue
-		}
-		if rerankIdentifierToken(field) {
-			out = append(out, field)
-			seen[field] = struct{}{}
-		}
-	}
-	return out
-}
-
-func rerankIdentifierToken(token string) bool {
-	if _, err := time.Parse("2006-01-02", token); err == nil {
-		return false
-	}
-	hasDigit := false
-	for _, r := range token {
-		if r >= '0' && r <= '9' {
-			hasDigit = true
-			break
-		}
-	}
-	return hasDigit && strings.Contains(token, "-")
-}
-
-func rerankText(value string) string {
-	value = strings.ToLower(strings.TrimSpace(value))
-	if value == "" {
-		return ""
-	}
-	replacer := strings.NewReplacer(
-		"\n", " ",
-		"\t", " ",
-		".", " ",
-		",", " ",
-		":", " ",
-		";", " ",
-		"?", " ",
-		"!", " ",
-		"(", " ",
-		")", " ",
-	)
-	return " " + strings.Join(strings.Fields(replacer.Replace(value)), " ") + " "
-}
-
-func containsAnyRerankCue(text string, cues []string) bool {
-	for _, cue := range cues {
-		if strings.Contains(text, cue) {
-			return true
-		}
-	}
-	return false
-}
-
-func containsAnyCue(text string, cues []string) bool {
-	return containsAnyRerankCue(text, cues)
-}
-
-func containsAnyAuthorityCue(text string, cues []string) bool {
-	return containsAnyRerankCue(text, cues)
-}
-
-var currentnessPositiveCues = []string{
-	" current ",
-	" now ",
-	" active ",
-	" valid on ",
-	" update dated ",
-}
-
-var currentnessStrongStaleCues = []string{
-	" archived ",
-	" obsolete ",
-	" replaced ",
-	" rejected ",
-	" not active ",
-	" not approved ",
-}
-
-var currentnessWeakStaleCues = []string{
-	" legacy ",
-	" previous ",
-	" previously ",
-	" before ",
-	" older ",
-	" old ",
-	" draft ",
-	" suggested ",
-	" copied ",
-	" rollback ",
-	" incident review ",
-	" proposed ",
-	" once ",
-	" future proposal ",
-}
-
-var directiveCues = []string{
-	" must use ",
-	" must be ",
-	" should use ",
-	" should be ",
-	" is assigned ",
-}
-
-var canonicalCues = []string{
-	" canonical ",
-	" current ",
-	" policy ",
-	" rule ",
-	" registry ",
-}
-
-var strongDisqualifierCues = []string{
-	" does not apply ",
-	" not about ",
-	" rejected ",
-	" unapproved ",
-	" forbidden ",
-	" false positive ",
-	" false positives ",
-}
-
-var weakDisqualifierCues = []string{
-	" legacy ",
-	" previously ",
-	" before ",
-	" once ",
-	" removed ",
-	" fallback ",
-	" draft ",
-	" rumor ",
-	" troubleshooting note ",
-}
-
-var conditionalQueryCues = []string{
-	" enterprise ",
-	" standard ",
-	" tenant ",
-	" exception ",
-}
-
-var authorityPositiveCues = []string{
-	" authoritative ",
-	" signed by ",
-	" approved ",
-	" official ",
-	" canonical ",
-	" source of truth ",
-}
-
-var authorityDirectiveCues = []string{
-	" requires ",
-	" require ",
-	" required ",
-	" must use ",
-	" must be ",
-}
-
-var authorityStrongNegativeCues = []string{
-	" informal chat ",
-	" not approved ",
-	" unapproved ",
-	" personal checklist ",
-	" meeting transcript ",
-}
-
-var authorityWeakNegativeCues = []string{
-	" suggested ",
-	" as an option ",
-	" while testing ",
-	" before ",
-	" draft ",
-	" note about ",
-}
-
-func factMatchesRecallWindow(f *domain.Fact, validAt, knownAt *time.Time) bool {
-	if f == nil {
-		return false
-	}
-	if validAt != nil {
-		if f.ValidFrom != nil && f.ValidFrom.After(*validAt) {
-			return false
-		}
-		if f.ValidTo != nil && !f.ValidTo.After(*validAt) {
-			return false
-		}
-	}
-	if knownAt != nil {
-		if f.RecordedAt.After(*knownAt) {
-			return false
-		}
-		if f.RecordedTo != nil && !f.RecordedTo.After(*knownAt) {
-			return false
-		}
-	}
-	return true
-}
-
-func claimMatchesRecallWindow(c *domain.Claim, validAt, knownAt *time.Time) bool {
-	if c == nil {
-		return false
-	}
-	if validAt != nil {
-		if c.ValidFrom != nil && c.ValidFrom.After(*validAt) {
-			return false
-		}
-		if c.ValidTo != nil && !c.ValidTo.After(*validAt) {
-			return false
-		}
-	}
-	if knownAt != nil {
-		if c.RecordedAt.After(*knownAt) {
-			return false
-		}
-		if c.RecordedTo != nil && !c.RecordedTo.After(*knownAt) {
-			return false
-		}
-	}
-	return true
-}
-
-func factCandidateMatchesRecallWindow(f FactRecallResult, validAt, knownAt *time.Time) bool {
-	if validAt != nil {
-		if f.ValidFrom != nil && f.ValidFrom.After(*validAt) {
-			return false
-		}
-		if f.ValidTo != nil && !f.ValidTo.After(*validAt) {
-			return false
-		}
-	}
-	if knownAt != nil {
-		if f.RecordedAt.After(*knownAt) {
-			return false
-		}
-		if f.RecordedTo != nil && !f.RecordedTo.After(*knownAt) {
-			return false
-		}
-	}
-	return true
-}
-
-func claimCandidateMatchesRecallWindow(c ClaimRecallResult, validAt, knownAt *time.Time) bool {
-	if validAt != nil {
-		if c.ValidFrom != nil && c.ValidFrom.After(*validAt) {
-			return false
-		}
-		if c.ValidTo != nil && !c.ValidTo.After(*validAt) {
-			return false
-		}
-	}
-	if knownAt != nil {
-		if c.RecordedAt.After(*knownAt) {
-			return false
-		}
-		if c.RecordedTo != nil && !c.RecordedTo.After(*knownAt) {
-			return false
-		}
-	}
-	return true
-}
-
-func filterSemanticFragmentsByWindow(hits []semanticsearch.SearchHit, validAt, knownAt *time.Time) []semanticsearch.SearchHit {
-	if validAt == nil && knownAt == nil {
-		return hits
-	}
-	out := hits[:0]
-	for _, hit := range hits {
-		if fragmentMetadataMatchesRecallWindow(hit.Metadata, validAt, knownAt) {
-			out = append(out, hit)
-		}
-	}
-	return out
-}
-
-func filterKeywordFragmentsByWindow(hits []keywordsearch.FragmentSearchResult, validAt, knownAt *time.Time) []keywordsearch.FragmentSearchResult {
-	if validAt == nil && knownAt == nil {
-		return hits
-	}
-	out := hits[:0]
-	for _, hit := range hits {
-		if fragmentMetadataMatchesRecallWindow(hit.Metadata, validAt, knownAt) {
-			out = append(out, hit)
-		}
-	}
-	return out
-}
-
-func fragmentMetadataMatchesRecallWindow(metadata map[string]any, validAt, knownAt *time.Time) bool {
-	if validAt != nil {
-		if validFrom, ok := metadataTime(metadata, "valid_from", "validFrom"); ok && validFrom.After(*validAt) {
-			return false
-		}
-		if validTo, ok := metadataTime(metadata, "valid_to", "validTo"); ok && !validTo.After(*validAt) {
-			return false
-		}
-	}
-	if knownAt != nil {
-		if recordedAt, ok := metadataTime(metadata, "recorded_at", "recordedAt"); ok && recordedAt.After(*knownAt) {
-			return false
-		}
-		if recordedTo, ok := metadataTime(metadata, "recorded_to", "recordedTo"); ok && !recordedTo.After(*knownAt) {
-			return false
-		}
-	}
-	return true
-}
-
-func metadataTime(metadata map[string]any, keys ...string) (time.Time, bool) {
-	for _, key := range keys {
-		value, ok := metadata[key]
-		if !ok {
-			continue
-		}
-		switch typed := value.(type) {
-		case time.Time:
-			return typed.UTC(), true
-		case string:
-			parsed, err := time.Parse(time.RFC3339Nano, typed)
-			if err != nil {
-				return time.Time{}, false
-			}
-			return parsed.UTC(), true
-		}
-	}
-	return time.Time{}, false
 }
 
 // filterSemanticFragments drops hits outside the caller's profile and any
