+++ internal/service/recallservice/recall_test.go
@@ -440,11 +440,41 @@ func TestRecallSmallHelpersAndNilLoggers(t *testing.T) {
 	require.Equal(t, "recall: embedding unavailable", sanitizeEmbeddingError(errors.New("other")).Error())
 	require.Equal(t, MaxLimit, clampLimit(MaxLimit+100))
 	require.Equal(t, DefaultLimit, clampLimit(0))
+	mergeRRFEntryTimes(nil, time.Now(), time.Now())
 
 	svc := &recallService{}
 	svc.logKeywordError(errors.New("keyword failed"))
 	svc.logHydrateError("fragment-1", errors.New("hydrate failed"))
 	svc.logEmbeddingError(errors.New("embedding failed"))
+}
+
+func TestBatchHydrateEvidenceFragmentsCoversBatchAndFallback(t *testing.T) {
+	ctx := context.Background()
+	svc := &recallService{}
+	evidenceSets := [][]domain.Evidence{
+		{{FragmentID: ""}, {FragmentID: "fragment-1"}, {FragmentID: "fragment-1"}},
+		{{FragmentID: "fragment-2"}},
+	}
+
+	require.Nil(t, svc.batchHydrateEvidenceFragments(ctx, "pA", evidenceSets))
+
+	batchHydrator := &fakeHydrator{}
+	svc.hydrator = batchHydrator
+	out := svc.batchHydrateEvidenceFragments(ctx, "pA", evidenceSets)
+	require.Len(t, out, 2)
+	require.Contains(t, out, "fragment-1")
+	require.Contains(t, out, "fragment-2")
+	require.Equal(t, int32(1), atomic.LoadInt32(&batchHydrator.batchCallCount))
+	require.Zero(t, atomic.LoadInt32(&batchHydrator.callCount))
+
+	require.Nil(t, svc.batchHydrateEvidenceFragments(ctx, "pA", [][]domain.Evidence{{{FragmentID: ""}}}))
+
+	fallbackHydrator := &fakeHydrator{batchErr: errors.New("batch failed")}
+	svc.hydrator = fallbackHydrator
+	out = svc.batchHydrateEvidenceFragments(ctx, "pA", evidenceSets)
+	require.Len(t, out, 2)
+	require.Equal(t, int32(1), atomic.LoadInt32(&fallbackHydrator.batchCallCount))
+	require.Equal(t, int32(2), atomic.LoadInt32(&fallbackHydrator.callCount))
 }
 
 // TestRecallService_OverfetchesVectorBranch — AC-40 overfetch requirement.
@@ -643,316 +673,6 @@ func TestRecallService_TierEnrichmentUsesQueryMatchedSearchers(t *testing.T) {
 	assert.Equal(t, int32(0), atomic.LoadInt32(&factGetter.callCount))
 	assert.Equal(t, int32(1), atomic.LoadInt32(&claimGetter.batchCallCount))
 	assert.Equal(t, int32(0), atomic.LoadInt32(&claimGetter.callCount))
-}
-
-func TestRecallService_CurrentnessRanksFactByTripleContentDate(t *testing.T) {
-	requiredRecordedAt := time.Date(2026, 6, 25, 12, 0, 0, 0, time.UTC)
-	negativeRecordedAt := time.Date(2026, 6, 29, 12, 0, 0, 0, time.UTC)
-	factSearcher := &fakeFactSearcher{
-		results: []FactRecallResult{
-			{FactID: "fact-required", ProfileID: "pA", RecordedAt: requiredRecordedAt},
-			{FactID: "fact-negative", ProfileID: "pA", RecordedAt: negativeRecordedAt},
-		},
-	}
-	factGetter := &fakeFactGetter{
-		facts: map[string]*domain.Fact{
-			"fact-required": {
-				FactID:     "fact-required",
-				ProfileID:  "pA",
-				Subject:    "service TIER-D-001 owner as of June 27, 2026",
-				Predicate:  "uses",
-				Object:     "owner-jade",
-				Status:     domain.FactStatusActive,
-				TruthScore: 0.7,
-				RecordedAt: requiredRecordedAt,
-			},
-			"fact-negative": {
-				FactID:     "fact-negative",
-				ProfileID:  "pA",
-				Subject:    "service TIER-D-001 owner as of June 20, 2026",
-				Predicate:  "uses",
-				Object:     "owner-onyx",
-				Status:     domain.FactStatusActive,
-				TruthScore: 0.99,
-				RecordedAt: negativeRecordedAt,
-			},
-		},
-	}
-	svc := NewRecallServiceWithTiers(
-		&stubEmbedding{DimensionsResult: 4},
-		&fakeSemanticSearcher{},
-		&fakeKeywordSearcher{},
-		&fakeHydrator{},
-		factSearcher,
-		factGetter,
-		nil,
-		nil,
-		0,
-		nil,
-		nil,
-	)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: "Who is the current owner for service TIER-D-001?", Limit: 1})
-
-	require.NoError(t, err)
-	require.Len(t, out, 1)
-	require.NotNil(t, out[0].Fact)
-	require.Equal(t, "fact-required", out[0].Fact.FactID)
-}
-
-func TestRecallService_CurrentnessRanksClaimByTripleContentDate(t *testing.T) {
-	requiredRecordedAt := time.Date(2026, 6, 25, 12, 0, 0, 0, time.UTC)
-	negativeRecordedAt := time.Date(2026, 6, 29, 12, 0, 0, 0, time.UTC)
-	claimSearcher := &fakeClaimSearcher{
-		results: []ClaimRecallResult{
-			{ClaimID: "claim-required", ProfileID: "pA", RecordedAt: requiredRecordedAt},
-			{ClaimID: "claim-negative", ProfileID: "pA", RecordedAt: negativeRecordedAt},
-		},
-	}
-	claimGetter := &fakeClaimGetter{
-		claims: map[string]*domain.Claim{
-			"claim-required": {
-				ClaimID:     "claim-required",
-				ProfileID:   "pA",
-				Subject:     "service TIER-D-002 pager as of June 27, 2026",
-				Predicate:   "uses",
-				Object:      "pager-jade",
-				Status:      domain.StatusValidated,
-				ExtractConf: 0.2,
-				RecordedAt:  requiredRecordedAt,
-			},
-			"claim-negative": {
-				ClaimID:     "claim-negative",
-				ProfileID:   "pA",
-				Subject:     "service TIER-D-002 pager as of June 20, 2026",
-				Predicate:   "uses",
-				Object:      "pager-onyx",
-				Status:      domain.StatusValidated,
-				ExtractConf: 0.99,
-				RecordedAt:  negativeRecordedAt,
-			},
-		},
-	}
-	svc := NewRecallServiceWithTiers(
-		&stubEmbedding{DimensionsResult: 4},
-		&fakeSemanticSearcher{},
-		&fakeKeywordSearcher{},
-		&fakeHydrator{},
-		nil,
-		nil,
-		claimSearcher,
-		claimGetter,
-		0,
-		nil,
-		nil,
-	)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: "What is the latest pager for service TIER-D-002?", Limit: 1})
-
-	require.NoError(t, err)
-	require.Len(t, out, 1)
-	require.NotNil(t, out[0].Claim)
-	require.Equal(t, "claim-required", out[0].Claim.ClaimID)
-}
-
-func TestRecallService_CurrentnessRanksFactByEvidenceFragmentDate(t *testing.T) {
-	requiredRecordedAt := time.Date(2026, 6, 25, 12, 0, 0, 0, time.UTC)
-	negativeRecordedAt := time.Date(2026, 6, 29, 12, 0, 0, 0, time.UTC)
-	factSearcher := &fakeFactSearcher{
-		results: []FactRecallResult{
-			{FactID: "fact-required", ProfileID: "pA", RecordedAt: requiredRecordedAt},
-			{FactID: "fact-negative", ProfileID: "pA", RecordedAt: negativeRecordedAt},
-		},
-	}
-	factGetter := &fakeFactGetter{
-		facts: map[string]*domain.Fact{
-			"fact-required": {
-				FactID:     "fact-required",
-				ProfileID:  "pA",
-				Subject:    "service TIER-S-001 owner",
-				Predicate:  "uses",
-				Object:     "owner-jade",
-				Status:     domain.FactStatusActive,
-				TruthScore: 0.7,
-				RecordedAt: requiredRecordedAt,
-				Evidence:   []domain.Evidence{{FragmentID: "fragment-required"}},
-			},
-			"fact-negative": {
-				FactID:     "fact-negative",
-				ProfileID:  "pA",
-				Subject:    "service TIER-S-001 owner",
-				Predicate:  "uses",
-				Object:     "owner-onyx",
-				Status:     domain.FactStatusActive,
-				TruthScore: 0.99,
-				RecordedAt: negativeRecordedAt,
-				Evidence:   []domain.Evidence{{FragmentID: "fragment-negative"}},
-			},
-		},
-	}
-	hydrator := &fakeHydrator{frags: map[string]*domain.Fragment{
-		"fragment-required": {
-			FragmentID: "fragment-required",
-			ProfileID:  "pA",
-			Content:    "Source update dated June 27, 2026. service TIER-S-001 owner uses owner-jade.",
-			CreatedAt:  requiredRecordedAt,
-			UpdatedAt:  requiredRecordedAt,
-		},
-		"fragment-negative": {
-			FragmentID: "fragment-negative",
-			ProfileID:  "pA",
-			Content:    "Source update dated June 20, 2026. service TIER-S-001 owner uses owner-onyx.",
-			CreatedAt:  negativeRecordedAt,
-			UpdatedAt:  negativeRecordedAt,
-		},
-	}}
-	svc := NewRecallServiceWithTiers(
-		&stubEmbedding{DimensionsResult: 4},
-		&fakeSemanticSearcher{},
-		&fakeKeywordSearcher{},
-		hydrator,
-		factSearcher,
-		factGetter,
-		nil,
-		nil,
-		0,
-		nil,
-		nil,
-	)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: "Who is the current owner for service TIER-S-001?", Limit: 1})
-
-	require.NoError(t, err)
-	require.Len(t, out, 1)
-	require.NotNil(t, out[0].Fact)
-	require.Equal(t, "fact-required", out[0].Fact.FactID)
-	require.Nil(t, out[0].Fact.Evidence)
-	require.Equal(t, int32(1), atomic.LoadInt32(&hydrator.batchCallCount))
-}
-
-func TestRecallService_CurrentnessRanksClaimByEvidenceFragmentDate(t *testing.T) {
-	requiredRecordedAt := time.Date(2026, 6, 25, 12, 0, 0, 0, time.UTC)
-	negativeRecordedAt := time.Date(2026, 6, 29, 12, 0, 0, 0, time.UTC)
-	claimSearcher := &fakeClaimSearcher{
-		results: []ClaimRecallResult{
-			{ClaimID: "claim-required", ProfileID: "pA", RecordedAt: requiredRecordedAt},
-			{ClaimID: "claim-negative", ProfileID: "pA", RecordedAt: negativeRecordedAt},
-		},
-	}
-	claimGetter := &fakeClaimGetter{
-		claims: map[string]*domain.Claim{
-			"claim-required": {
-				ClaimID:     "claim-required",
-				ProfileID:   "pA",
-				Subject:     "service TIER-S-002 pager",
-				Predicate:   "uses",
-				Object:      "pager-jade",
-				Status:      domain.StatusValidated,
-				ExtractConf: 0.2,
-				RecordedAt:  requiredRecordedAt,
-				Evidence:    []domain.Evidence{{FragmentID: "fragment-required"}},
-			},
-			"claim-negative": {
-				ClaimID:     "claim-negative",
-				ProfileID:   "pA",
-				Subject:     "service TIER-S-002 pager",
-				Predicate:   "uses",
-				Object:      "pager-onyx",
-				Status:      domain.StatusValidated,
-				ExtractConf: 0.99,
-				RecordedAt:  negativeRecordedAt,
-				Evidence:    []domain.Evidence{{FragmentID: "fragment-negative"}},
-			},
-		},
-	}
-	hydrator := &fakeHydrator{frags: map[string]*domain.Fragment{
-		"fragment-required": {
-			FragmentID: "fragment-required",
-			ProfileID:  "pA",
-			Content:    "Source update dated June 27, 2026. service TIER-S-002 pager uses pager-jade.",
-			CreatedAt:  requiredRecordedAt,
-			UpdatedAt:  requiredRecordedAt,
-		},
-		"fragment-negative": {
-			FragmentID: "fragment-negative",
-			ProfileID:  "pA",
-			Content:    "Source update dated June 20, 2026. service TIER-S-002 pager uses pager-onyx.",
-			CreatedAt:  negativeRecordedAt,
-			UpdatedAt:  negativeRecordedAt,
-		},
-	}}
-	svc := NewRecallServiceWithTiers(
-		&stubEmbedding{DimensionsResult: 4},
-		&fakeSemanticSearcher{},
-		&fakeKeywordSearcher{},
-		hydrator,
-		nil,
-		nil,
-		claimSearcher,
-		claimGetter,
-		0,
-		nil,
-		nil,
-	)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: "What is the latest pager for service TIER-S-002?", Limit: 1})
-
-	require.NoError(t, err)
-	require.Len(t, out, 1)
-	require.NotNil(t, out[0].Claim)
-	require.Equal(t, "claim-required", out[0].Claim.ClaimID)
-	require.Nil(t, out[0].Claim.Evidence)
-	require.Equal(t, int32(1), atomic.LoadInt32(&hydrator.batchCallCount))
-}
-
-func TestTemporalRankTimeForRecallPrefersValidFromOverTripleContentDate(t *testing.T) {
-	validFrom := time.Date(2026, 6, 20, 0, 0, 0, 0, time.UTC)
-	recordedAt := time.Date(2026, 6, 29, 12, 0, 0, 0, time.UTC)
-
-	got := temporalRankTimeForRecall(
-		"Who is the current owner for service TIER-D-003?",
-		&validFrom,
-		recordedAt,
-		"service TIER-D-003 owner as of June 27, 2026",
-		"uses",
-		"owner-jade",
-	)
-
-	require.Equal(t, validFrom, got)
-}
-
-func TestLatestTemporalDateInTextParsesWeekdayNames(t *testing.T) {
-	anchor := time.Date(2026, 6, 29, 12, 0, 0, 0, time.UTC)
-
-	require.Equal(
-		t,
-		time.Date(2026, 6, 26, 0, 0, 0, 0, time.UTC),
-		latestTemporalDateInText("Owner update from Friday.", anchor),
-	)
-	require.Equal(
-		t,
-		time.Date(2026, 6, 26, 0, 0, 0, 0, time.UTC),
-		latestTemporalDateInText("Owner update from last Friday.", anchor),
-	)
-	require.Equal(
-		t,
-		time.Date(2026, 6, 22, 0, 0, 0, 0, time.UTC),
-		latestTemporalDateInText("Owner update from last Monday.", anchor),
-	)
-	require.Zero(t, latestTemporalDateInText("Owner update expected next Friday.", anchor))
-}
-
-func TestLatestTemporalDateInTextParsesUnambiguousNumericDates(t *testing.T) {
-	anchor := time.Date(2026, 6, 29, 12, 0, 0, 0, time.UTC)
-	expected := time.Date(2026, 6, 27, 0, 0, 0, 0, time.UTC)
-
-	require.Equal(t, expected, latestTemporalDateInText("Owner update dated 6/27/2026.", anchor))
-	require.Equal(t, expected, latestTemporalDateInText("Owner update dated 27/6/2026.", anchor))
-	require.Equal(t, expected, latestTemporalDateInText("Owner update dated 2026/6/27.", anchor))
-	require.Equal(t, expected, latestTemporalDateInText("Owner update dated 6/27.", anchor))
-	require.Zero(t, latestTemporalDateInText("Owner update dated 6/7/2026.", anchor))
-	require.Zero(t, latestTemporalDateInText("Owner update dated 2/31/2026.", anchor))
-	require.Zero(t, latestTemporalDateInText("Owner update dated 6/27.", time.Time{}))
 }
 
 func TestRecallService_EvidenceSourceQueryPrefersFragmentOverDerivedFact(t *testing.T) {
@@ -1189,803 +909,6 @@ func TestRecallService_RRFScoreOrdering(t *testing.T) {
 	if topScore <= midScore {
 		t.Errorf("top score %v must exceed single-branch score %v", topScore, midScore)
 	}
-}
-
-func TestRecallService_CurrentnessRerankPrefersCurrentUpdate(t *testing.T) {
-	query := "What is the current deployment window for service OBS-001?"
-	sem := &fakeSemanticSearcher{
-		hits: []semanticsearch.SearchHit{
-			{
-				ID:      "f-archived",
-				Type:    "fragment",
-				Content: "Archived 2025 calendar. service OBS-001 deployed at 02:01 UTC before the correction.",
-			},
-		},
-	}
-	kw := &fakeKeywordSearcher{
-		hits: []keywordsearch.FragmentSearchResult{
-			{
-				FragmentID: "f-current",
-				Content:    "Current release calendar update dated 2026-06-28. service OBS-001 now deploys at 03:01 UTC.",
-			},
-			{
-				FragmentID: "f-archived",
-				Content:    "Archived 2025 calendar. service OBS-001 deployed at 02:01 UTC before the correction.",
-			},
-		},
-	}
-	emb := &stubEmbedding{DimensionsResult: 4}
-	svc := NewRecallService(emb, sem, kw, &fakeHydrator{}, nil, nil)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: query, Limit: 5})
-
-	require.NoError(t, err)
-	require.NotEmpty(t, out)
-	require.Equal(t, "f-current", out[0].Fragment.FragmentID)
-}
-
-func TestCurrentnessAdjustmentRequiresQueryIdentifiersForBoosts(t *testing.T) {
-	query := "What is the current deployment window for service OBS-001?"
-	neighbor := "Current release calendar update dated 2026-06-28. service OBS-002 now deploys at 03:02 UTC."
-
-	require.Zero(t, currentnessAdjustment(query, neighbor))
-}
-
-func TestRecallService_CurrentnessRerankPrefersNewerDatedFragment(t *testing.T) {
-	query := "Who is the current owner for account TMP-001?"
-	oldCreated := time.Date(2026, 6, 19, 12, 0, 0, 0, time.UTC)
-	newCreated := time.Date(2026, 6, 24, 12, 0, 0, 0, time.UTC)
-	sem := &fakeSemanticSearcher{
-		hits: []semanticsearch.SearchHit{
-			{
-				ID:        "f-old-current",
-				Type:      "fragment",
-				Content:   "Current owner record. account TMP-001 owner is alice.",
-				CreatedAt: oldCreated,
-				UpdatedAt: oldCreated,
-			},
-		},
-	}
-	kw := &fakeKeywordSearcher{
-		hits: []keywordsearch.FragmentSearchResult{
-			{
-				FragmentID: "f-new-dated",
-				Content:    "Owner record dated 2026-06-26. account TMP-001 owner is bob.",
-				CreatedAt:  newCreated,
-				UpdatedAt:  newCreated,
-			},
-		},
-	}
-	emb := &stubEmbedding{DimensionsResult: 4}
-	svc := NewRecallService(emb, sem, kw, &fakeHydrator{}, nil, nil)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: query, Limit: 5})
-
-	require.NoError(t, err)
-	require.NotEmpty(t, out)
-	require.Equal(t, "f-new-dated", out[0].Fragment.FragmentID)
-}
-
-func TestRecallService_ValidAtFiltersFutureFragmentMetadata(t *testing.T) {
-	query := "As of 2026-06-22, what owner did service TIER-W-001 use?"
-	validAt := time.Date(2026, 6, 22, 0, 0, 0, 0, time.UTC)
-	sem := &fakeSemanticSearcher{
-		hits: []semanticsearch.SearchHit{
-			{
-				ID:       "f-future",
-				Type:     "fragment",
-				Content:  "Since 2026-06-27, service TIER-W-001 owner uses owner-cobalt.",
-				Metadata: map[string]any{"valid_from": "2026-06-27T00:00:00Z"},
-			},
-			{
-				ID:       "f-old",
-				Type:     "fragment",
-				Content:  "Since 2026-06-20, service TIER-W-001 owner uses owner-coral.",
-				Metadata: map[string]any{"valid_from": "2026-06-20T00:00:00Z"},
-			},
-		},
-	}
-	kw := &fakeKeywordSearcher{
-		hits: []keywordsearch.FragmentSearchResult{
-			{
-				FragmentID: "f-future",
-				Content:    "Since 2026-06-27, service TIER-W-001 owner uses owner-cobalt.",
-				Metadata:   map[string]any{"valid_from": "2026-06-27T00:00:00Z"},
-			},
-			{
-				FragmentID: "f-old",
-				Content:    "Since 2026-06-20, service TIER-W-001 owner uses owner-coral.",
-				Metadata:   map[string]any{"valid_from": "2026-06-20T00:00:00Z"},
-			},
-		},
-	}
-	emb := &stubEmbedding{DimensionsResult: 4}
-	svc := NewRecallService(emb, sem, kw, &fakeHydrator{}, nil, nil)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: query, Limit: 5, ValidAt: &validAt})
-
-	require.NoError(t, err)
-	require.NotEmpty(t, out)
-	require.Equal(t, "f-old", out[0].Fragment.FragmentID)
-	for _, hit := range out {
-		if hit.Fragment != nil {
-			require.NotEqual(t, "f-future", hit.Fragment.FragmentID)
-		}
-	}
-}
-
-func TestFragmentMetadataKnownAtRequiresExplicitRecordedMetadata(t *testing.T) {
-	knownAt := time.Date(2026, 6, 28, 0, 0, 0, 0, time.UTC)
-
-	require.True(t, fragmentMetadataMatchesRecallWindow(nil, nil, &knownAt))
-	require.False(t, fragmentMetadataMatchesRecallWindow(
-		map[string]any{"recorded_at": "2026-06-29T00:00:00Z"},
-		nil,
-		&knownAt,
-	))
-}
-
-func TestRecallService_CurrentnessRerankDatedFragmentBeatsUndatedCurrentCue(t *testing.T) {
-	query := "Who is the current owner for account TMP-001?"
-	oldCreated := time.Date(2026, 6, 19, 12, 0, 0, 0, time.UTC)
-	newCreated := time.Date(2026, 6, 24, 12, 0, 0, 0, time.UTC)
-	oldCurrent := semanticsearch.SearchHit{
-		ID:        "f-old-current",
-		Type:      "fragment",
-		Content:   "Current owner record. account TMP-001 owner is alice.",
-		CreatedAt: oldCreated,
-		UpdatedAt: oldCreated,
-	}
-	sem := &fakeSemanticSearcher{
-		hits: []semanticsearch.SearchHit{oldCurrent},
-	}
-	kw := &fakeKeywordSearcher{
-		hits: []keywordsearch.FragmentSearchResult{
-			{
-				FragmentID: "f-old-current",
-				Content:    oldCurrent.Content,
-				CreatedAt:  oldCreated,
-				UpdatedAt:  oldCreated,
-			},
-			{
-				FragmentID: "f-new-dated",
-				Content:    "Owner record dated 2026-06-26. account TMP-001 owner is bob.",
-				CreatedAt:  newCreated,
-				UpdatedAt:  newCreated,
-			},
-		},
-	}
-	svc := NewRecallService(&stubEmbedding{DimensionsResult: 4}, sem, kw, &fakeHydrator{}, nil, nil)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: query, Limit: 5})
-
-	require.NoError(t, err)
-	require.NotEmpty(t, out)
-	require.Equal(t, "f-new-dated", out[0].Fragment.FragmentID)
-}
-
-func TestRecallService_CurrentnessRerankPrefersRelativeDatedFragment(t *testing.T) {
-	query := "Who is the current owner for service TMP-002?"
-	importedAt := time.Date(2026, 6, 28, 12, 0, 0, 0, time.UTC)
-	sem := &fakeSemanticSearcher{
-		hits: []semanticsearch.SearchHit{
-			{
-				ID:        "f-undated-current",
-				Type:      "fragment",
-				Content:   "Current owner note. service TMP-002 owner uses owner-moon.",
-				CreatedAt: importedAt,
-				UpdatedAt: importedAt,
-			},
-		},
-	}
-	kw := &fakeKeywordSearcher{
-		hits: []keywordsearch.FragmentSearchResult{
-			{
-				FragmentID: "f-relative-dated",
-				Content:    "Owner update from yesterday. service TMP-002 owner uses owner-sun.",
-				CreatedAt:  importedAt,
-				UpdatedAt:  importedAt,
-			},
-			{
-				FragmentID: "f-undated-current",
-				Content:    "Current owner note. service TMP-002 owner uses owner-moon.",
-				CreatedAt:  importedAt,
-				UpdatedAt:  importedAt,
-			},
-		},
-	}
-	svc := NewRecallService(&stubEmbedding{DimensionsResult: 4}, sem, kw, &fakeHydrator{}, nil, nil)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: query, Limit: 5})
-
-	require.NoError(t, err)
-	require.NotEmpty(t, out)
-	require.Equal(t, "f-relative-dated", out[0].Fragment.FragmentID)
-}
-
-func TestRecallService_CurrentnessRerankPrefersMonthNameDatedFragment(t *testing.T) {
-	query := "Who is the current owner for service TMP-003?"
-	importedAt := time.Date(2026, 6, 28, 12, 0, 0, 0, time.UTC)
-	sem := &fakeSemanticSearcher{
-		hits: []semanticsearch.SearchHit{
-			{
-				ID:        "f-undated-current",
-				Type:      "fragment",
-				Content:   "Current owner note. service TMP-003 owner uses owner-moon.",
-				CreatedAt: importedAt,
-				UpdatedAt: importedAt,
-			},
-		},
-	}
-	kw := &fakeKeywordSearcher{
-		hits: []keywordsearch.FragmentSearchResult{
-			{
-				FragmentID: "f-month-dated",
-				Content:    "Owner update dated June 27, 2026. service TMP-003 owner uses owner-sun.",
-				CreatedAt:  importedAt,
-				UpdatedAt:  importedAt,
-			},
-			{
-				FragmentID: "f-undated-current",
-				Content:    "Current owner note. service TMP-003 owner uses owner-moon.",
-				CreatedAt:  importedAt,
-				UpdatedAt:  importedAt,
-			},
-		},
-	}
-	svc := NewRecallService(&stubEmbedding{DimensionsResult: 4}, sem, kw, &fakeHydrator{}, nil, nil)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: query, Limit: 5})
-
-	require.NoError(t, err)
-	require.NotEmpty(t, out)
-	require.Equal(t, "f-month-dated", out[0].Fragment.FragmentID)
-}
-
-func TestRecallService_CurrentnessRerankDemotesExpiredValidityRange(t *testing.T) {
-	query := "Who is the current owner for service TMP-004?"
-	importedAt := time.Date(2026, 6, 29, 12, 0, 0, 0, time.UTC)
-	expired := "Current owner note from the old rollout. service TMP-004 owner uses owner-cinder. This assignment is valid through 2026-06-20."
-	current := "Current owner note. service TMP-004 owner uses owner-aurora."
-	sem := &fakeSemanticSearcher{
-		hits: []semanticsearch.SearchHit{
-			{
-				ID:        "f-expired-range",
-				Type:      "fragment",
-				Content:   expired,
-				CreatedAt: importedAt,
-				UpdatedAt: importedAt,
-			},
-		},
-	}
-	kw := &fakeKeywordSearcher{
-		hits: []keywordsearch.FragmentSearchResult{
-			{
-				FragmentID: "f-expired-range",
-				Content:    expired,
-				CreatedAt:  importedAt,
-				UpdatedAt:  importedAt,
-			},
-			{
-				FragmentID: "f-current",
-				Content:    current,
-				CreatedAt:  importedAt,
-				UpdatedAt:  importedAt,
-			},
-		},
-	}
-	svc := NewRecallService(&stubEmbedding{DimensionsResult: 4}, sem, kw, &fakeHydrator{}, nil, nil)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: query, Limit: 5})
-
-	require.NoError(t, err)
-	require.NotEmpty(t, out)
-	require.Equal(t, "f-current", out[0].Fragment.FragmentID)
-}
-
-func TestRecallService_CurrentnessRerankHandlesActiveDecisionQueries(t *testing.T) {
-	query := "What is the active launch decision for notice RET-046?"
-	required := "Retraction update dated 2026-06-28. Earlier approval for notice RET-046 was withdrawn. The active decision is launch-paused-046."
-	stale := "Original announcement said notice RET-046 was launch-approved-046 before the retraction."
-	sem := &fakeSemanticSearcher{
-		hits: []semanticsearch.SearchHit{
-			{ID: "f-stale", Type: "fragment", Content: stale},
-			{ID: "f-required", Type: "fragment", Content: required},
-		},
-	}
-	kw := &fakeKeywordSearcher{
-		hits: []keywordsearch.FragmentSearchResult{
-			{FragmentID: "f-required", Content: required},
-			{FragmentID: "f-stale", Content: stale},
-		},
-	}
-	svc := NewRecallService(&stubEmbedding{DimensionsResult: 4}, sem, kw, &fakeHydrator{}, nil, nil)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: query, Limit: 5})
-
-	require.NoError(t, err)
-	require.NotEmpty(t, out)
-	require.Equal(t, "f-required", out[0].Fragment.FragmentID)
-}
-
-func TestCurrentnessTemporalAdjustmentRequiresQueryIdentifiers(t *testing.T) {
-	query := "Who is the current owner for account TMP-001?"
-	newest := time.Date(2026, 6, 26, 0, 0, 0, 0, time.UTC)
-	frame := currentnessTemporalFrame{hasContentDate: true, newestContentDate: newest}
-	neighbor := rrfEntry{
-		Content:   "Owner record dated 2026-06-26. account TMP-002 owner is bob.",
-		CreatedAt: newest,
-		UpdatedAt: newest,
-	}
-
-	require.Zero(t, currentnessTemporalAdjustment(query, neighbor, frame))
-}
-
-func TestCurrentnessTemporalFramePrefersContentDatesOverBulkImportTimes(t *testing.T) {
-	query := "What is the current deployment window for service OBS-001?"
-	importedAt := time.Date(2026, 6, 28, 20, 0, 0, 0, time.UTC)
-	entries := []rrfEntry{
-		{
-			Content:   "Current release calendar update dated 2026-06-28. service OBS-001 now deploys at 03:01 UTC.",
-			CreatedAt: importedAt,
-			UpdatedAt: importedAt,
-		},
-		{
-			Content:   "Draft migration plan suggested 02:01 UTC for service OBS-001, but the plan was replaced.",
-			CreatedAt: importedAt,
-			UpdatedAt: importedAt,
-		},
-	}
-
-	frame := currentnessTemporalFrameFor(query, entries)
-
-	require.True(t, frame.hasContentDate)
-	require.False(t, frame.useFragmentTimestamp)
-	require.Positive(t, currentnessTemporalAdjustment(query, entries[0], frame))
-	require.Negative(t, currentnessTemporalAdjustment(query, entries[1], frame))
-}
-
-func TestExpiredValidityAdjustmentUsesAsOfTime(t *testing.T) {
-	query := "Who is the current owner for service TMP-004?"
-	entry := rrfEntry{
-		Content:   "Current owner note. service TMP-004 owner uses owner-cinder. This assignment is valid through June 30, 2026.",
-		CreatedAt: time.Date(2026, 6, 29, 12, 0, 0, 0, time.UTC),
-		UpdatedAt: time.Date(2026, 6, 29, 12, 0, 0, 0, time.UTC),
-	}
-	frame := currentnessTemporalFrame{newestFragmentTime: entry.CreatedAt}
-
-	require.Zero(t, expiredValidityAdjustment(query, entry, frame))
-
-	frame.newestFragmentTime = time.Date(2026, 7, 1, 12, 0, 0, 0, time.UTC)
-
-	require.Negative(t, expiredValidityAdjustment(query, entry, frame))
-}
-
-func TestLatestTemporalDateInEntryParsesRelativeDatesFromFragmentTimestamp(t *testing.T) {
-	anchor := time.Date(2026, 6, 28, 18, 30, 0, 0, time.UTC)
-	cases := []struct {
-		name    string
-		content string
-		want    time.Time
-	}{
-		{
-			name:    "yesterday",
-			content: "Owner update from yesterday. service TMP-002 owner uses owner-sun.",
-			want:    time.Date(2026, 6, 27, 0, 0, 0, 0, time.UTC),
-		},
-		{
-			name:    "numeric days ago",
-			content: "Owner update from 2 days ago. service TMP-002 owner uses owner-sun.",
-			want:    time.Date(2026, 6, 26, 0, 0, 0, 0, time.UTC),
-		},
-		{
-			name:    "word days ago",
-			content: "Owner update from two days ago. service TMP-002 owner uses owner-sun.",
-			want:    time.Date(2026, 6, 26, 0, 0, 0, 0, time.UTC),
-		},
-		{
-			name:    "last week",
-			content: "Owner update from last week. service TMP-002 owner uses owner-sun.",
-			want:    time.Date(2026, 6, 21, 0, 0, 0, 0, time.UTC),
-		},
-		{
-			name:    "month name with year",
-			content: "Owner update dated June 27, 2026. service TMP-002 owner uses owner-sun.",
-			want:    time.Date(2026, 6, 27, 0, 0, 0, 0, time.UTC),
-		},
-		{
-			name:    "abbreviated month with anchored year",
-			content: "Owner update dated Jun 27. service TMP-002 owner uses owner-sun.",
-			want:    time.Date(2026, 6, 27, 0, 0, 0, 0, time.UTC),
-		},
-		{
-			name:    "day before month",
-			content: "Owner update dated 27 June 2026. service TMP-002 owner uses owner-sun.",
-			want:    time.Date(2026, 6, 27, 0, 0, 0, 0, time.UTC),
-		},
-	}
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			got := latestTemporalDateInEntry(rrfEntry{Content: tc.content, CreatedAt: anchor, UpdatedAt: anchor})
-
-			require.Equal(t, tc.want, got)
-		})
-	}
-}
-
-func TestLatestCurrentnessTemporalDateInEntryIgnoresValidityEndOnlyDate(t *testing.T) {
-	anchor := time.Date(2026, 6, 29, 12, 0, 0, 0, time.UTC)
-	entry := rrfEntry{
-		Content:   "Current owner note. service TMP-004 owner uses owner-cinder. This assignment is valid through 2026-06-20.",
-		CreatedAt: anchor,
-		UpdatedAt: anchor,
-	}
-
-	require.Equal(t, time.Date(2026, 6, 20, 0, 0, 0, 0, time.UTC), latestValidityEndDateInEntry(entry))
-	require.Zero(t, latestCurrentnessTemporalDateInEntry(entry))
-
-	entry.Content = "Owner update dated 2026-06-27. service TMP-004 owner uses owner-aurora; the old assignment was valid through 2026-06-20."
-	require.Equal(t, time.Date(2026, 6, 27, 0, 0, 0, 0, time.UTC), latestCurrentnessTemporalDateInEntry(entry))
-}
-
-func TestFilterNonPositiveRRFEntriesDropsZeroScoresWhenPositiveExists(t *testing.T) {
-	entries := []rrfEntry{
-		{id: "positive", FinalScore: 0.01},
-		{id: "zero", FinalScore: 0},
-	}
-
-	out := filterNonPositiveRRFEntries(entries)
-
-	require.Len(t, out, 1)
-	require.Equal(t, "positive", out[0].id)
-}
-
-func TestFilterNonPositiveRRFEntriesKeepsAllZeroScoresWhenNoPositiveExists(t *testing.T) {
-	entries := []rrfEntry{
-		{id: "zero-1", FinalScore: 0},
-		{id: "zero-2", FinalScore: 0},
-	}
-
-	out := filterNonPositiveRRFEntries(entries)
-
-	require.Len(t, out, 2)
-	require.Equal(t, "zero-1", out[0].id)
-	require.Equal(t, "zero-2", out[1].id)
-}
-
-func TestRecallService_IdentifierSpecificityPrefersExactJobID(t *testing.T) {
-	query := "What timeout should job UNT-013 use?"
-	neighbor := "Runtime configuration. job UNT-003 must use a timeout of 13 minutes."
-	required := "Runtime configuration. job UNT-013 must use a timeout of 23 minutes."
-	sem := &fakeSemanticSearcher{
-		hits: []semanticsearch.SearchHit{
-			{ID: "f-neighbor", Type: "fragment", Content: neighbor},
-			{ID: "f-filler-1", Type: "fragment", Content: "Runtime configuration. job UNT-001 must use a timeout of 11 minutes."},
-			{ID: "f-filler-2", Type: "fragment", Content: "Runtime configuration. job UNT-041 must use a timeout of 11 minutes."},
-			{ID: "f-required", Type: "fragment", Content: required},
-		},
-	}
-	kw := &fakeKeywordSearcher{
-		hits: []keywordsearch.FragmentSearchResult{
-			{FragmentID: "f-required", Content: required},
-			{FragmentID: "f-filler-1", Content: "Runtime configuration. job UNT-001 must use a timeout of 11 minutes."},
-			{FragmentID: "f-neighbor", Content: neighbor},
-		},
-	}
-	svc := NewRecallService(&stubEmbedding{DimensionsResult: 4}, sem, kw, &fakeHydrator{}, nil, nil)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: query, Limit: 5})
-
-	require.NoError(t, err)
-	require.NotEmpty(t, out)
-	require.Equal(t, "f-required", out[0].Fragment.FragmentID)
-}
-
-func TestIdentifierSpecificityAdjustmentRequiresExactIdentifier(t *testing.T) {
-	queryText := rerankText("What timeout should job UNT-013 use?")
-
-	require.Positive(t, identifierSpecificityAdjustment(queryText, "Runtime configuration. job UNT-013 must use a timeout of 23 minutes."))
-	require.Zero(t, identifierSpecificityAdjustment(queryText, "Runtime configuration. job UNT-003 must use a timeout of 13 minutes."))
-}
-
-func TestRerankIdentifiersSkipsISODateTokens(t *testing.T) {
-	queryText := rerankText("As of 2026-06-22, what owner did service TIER-W-001 use?")
-
-	require.Equal(t, []string{"tier-w-001"}, rerankIdentifiers(queryText))
-}
-
-func TestApplyIdentifierSpecificityAdjustmentsRequiresUnitValueQuery(t *testing.T) {
-	entries := []rrfEntry{
-		{
-			Content:    "Current release calendar update dated 2026-06-28. service OBS-001 now deploys at 03:01 UTC.",
-			FinalScore: 0.02,
-		},
-	}
-
-	applyIdentifierSpecificityAdjustments("What is the current deployment window for service OBS-001?", entries)
-
-	require.Equal(t, 0.02, entries[0].FinalScore)
-}
-
-func TestRecallService_CueRerankPrefersDirectiveEvidence(t *testing.T) {
-	query := "Which pager should alerts for queue NEG-001 use?"
-	sem := &fakeSemanticSearcher{
-		hits: []semanticsearch.SearchHit{
-			{
-				ID:      "f-stale",
-				Type:    "fragment",
-				Content: "Before 2026, queue NEG-001 used pager-001-red for alerts.",
-			},
-		},
-	}
-	kw := &fakeKeywordSearcher{
-		hits: []keywordsearch.FragmentSearchResult{
-			{
-				FragmentID: "f-required",
-				Content:    "Routing policy dated 2026-06-28. Alerts for queue NEG-001 must use pager-001-green.",
-			},
-			{
-				FragmentID: "f-stale",
-				Content:    "Before 2026, queue NEG-001 used pager-001-red for alerts.",
-			},
-		},
-	}
-	emb := &stubEmbedding{DimensionsResult: 4}
-	svc := NewRecallService(emb, sem, kw, &fakeHydrator{}, nil, nil)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: query, Limit: 5})
-
-	require.NoError(t, err)
-	require.NotEmpty(t, out)
-	require.Equal(t, "f-required", out[0].Fragment.FragmentID)
-	for _, hit := range out {
-		require.NotEqual(t, "f-stale", hit.Fragment.FragmentID)
-	}
-}
-
-func TestRecallService_CueRerankSuppressesHistoricalSiblingWhenDirectiveExists(t *testing.T) {
-	query := "Which pager should alerts for queue NEG-001 use?"
-	stale := "Before 2026, queue NEG-001 used pager-001-red for alerts."
-	sem := &fakeSemanticSearcher{
-		hits: []semanticsearch.SearchHit{
-			{ID: "f-stale", Type: "fragment", Content: stale},
-		},
-	}
-	kw := &fakeKeywordSearcher{
-		hits: []keywordsearch.FragmentSearchResult{
-			{
-				FragmentID: "f-required",
-				Content:    "Routing policy dated 2026-06-28. Alerts for queue NEG-001 must use pager-001-green. Do not use pager-001-red for this queue.",
-			},
-			{FragmentID: "f-stale", Content: stale},
-			{FragmentID: "f-filler-1", Content: "Queue NEG-001 inventory reference for alert routing review."},
-			{FragmentID: "f-filler-2", Content: "Queue NEG-001 runbook index entry for pager ownership audits."},
-			{FragmentID: "f-filler-3", Content: "Queue NEG-001 on-call calendar note for escalation metadata."},
-			{FragmentID: "f-filler-4", Content: "Queue NEG-001 dashboard bookmark for support handoff."},
-		},
-	}
-	svc := NewRecallService(&stubEmbedding{DimensionsResult: 4}, sem, kw, &fakeHydrator{}, nil, nil)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: query, Limit: 5})
-
-	require.NoError(t, err)
-	require.Len(t, out, 5)
-	for _, hit := range out {
-		require.NotEqual(t, "f-stale", hit.Fragment.FragmentID)
-	}
-}
-
-func TestHistoricalSelectionAdjustmentRequiresDirectiveFrame(t *testing.T) {
-	query := "Which pager should alerts for queue NEG-001 use?"
-	stale := "Before 2026, queue NEG-001 used pager-001-red for alerts."
-
-	require.Zero(t, historicalSelectionAdjustment(query, stale, selectionCueFrame{}))
-	require.Negative(t, historicalSelectionAdjustment(query, stale, selectionCueFrame{hasDirectiveMatch: true}))
-}
-
-func TestRecallCueAdjustmentRequiresQueryIdentifiersForBoosts(t *testing.T) {
-	query := "Which endpoint should the west-030 region use for billing sync?"
-	neighborTemplate := "Export routing rule. Enterprise tenants such as tenant CND-030 enterprise must use endpoint-enterprise-030."
-
-	require.Zero(t, cueAdjustment(query, neighborTemplate))
-}
-
-func TestRecallService_AuthorityRerankPrefersSignedRunbook(t *testing.T) {
-	query := "Which procedure does runbook AUT-061 require?"
-	sem := &fakeSemanticSearcher{
-		hits: []semanticsearch.SearchHit{
-			{
-				ID:      "f-chat",
-				Type:    "fragment",
-				Content: "An informal chat suggested procedure-061-chat for runbook AUT-061, but it was not approved.",
-			},
-		},
-	}
-	kw := &fakeKeywordSearcher{
-		hits: []keywordsearch.FragmentSearchResult{
-			{
-				FragmentID: "f-required",
-				Content:    "Authoritative runbook signed by operations. runbook AUT-061 requires procedure-061-canonical.",
-			},
-			{
-				FragmentID: "f-chat",
-				Content:    "An informal chat suggested procedure-061-chat for runbook AUT-061, but it was not approved.",
-			},
-		},
-	}
-	emb := &stubEmbedding{DimensionsResult: 4}
-	svc := NewRecallService(emb, sem, kw, &fakeHydrator{}, nil, nil)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: query, Limit: 5})
-
-	require.NoError(t, err)
-	require.NotEmpty(t, out)
-	require.Equal(t, "f-required", out[0].Fragment.FragmentID)
-}
-
-func TestAuthorityAdjustmentRequiresQueryIdentifiersForBoosts(t *testing.T) {
-	query := "Which procedure does runbook AUT-061 require?"
-	neighbor := "Authoritative runbook signed by operations. runbook AUT-062 requires procedure-062-canonical."
-
-	require.Zero(t, authorityAdjustment(query, neighbor))
-}
-
-func TestRecallService_CommunityExpansionDisabledByDefault(t *testing.T) {
-	sem := &fakeSemanticSearcher{hits: []semanticsearch.SearchHit{{ID: "f-direct", Type: "fragment"}}}
-	kw := &fakeKeywordSearcher{}
-	expander := &fakeCommunityExpander{
-		expansion: CommunityExpansion{
-			SelectedCommunities: 1,
-			Fragments: []CommunityFragmentRecallResult{
-				{FragmentID: "f-community", ProfileID: "pA", Score: 0.8},
-			},
-		},
-	}
-	svc := NewRecallServiceWithTiers(
-		&stubEmbedding{DimensionsResult: 4},
-		sem,
-		kw,
-		&fakeHydrator{},
-		nil,
-		nil,
-		nil,
-		nil,
-		0,
-		nil,
-		nil,
-		WithCommunityExpander(expander),
-	)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: "q", Limit: 5})
-	require.NoError(t, err)
-	require.Len(t, out, 1)
-	require.Equal(t, "f-direct", out[0].Fragment.FragmentID)
-	require.Equal(t, 0, expander.calls)
-}
-
-func TestRecallService_CommunityExpansionNoCommunityFallback(t *testing.T) {
-	sem := &fakeSemanticSearcher{hits: []semanticsearch.SearchHit{{ID: "f-direct", Type: "fragment"}}}
-	kw := &fakeKeywordSearcher{}
-	expander := &fakeCommunityExpander{}
-	svc := NewRecallServiceWithTiers(
-		&stubEmbedding{DimensionsResult: 4},
-		sem,
-		kw,
-		&fakeHydrator{},
-		nil,
-		nil,
-		nil,
-		nil,
-		0,
-		nil,
-		nil,
-		WithCommunityExpander(expander),
-	)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: "q", Limit: 5, UseCommunities: true})
-	require.NoError(t, err)
-	require.Len(t, out, 1)
-	require.Equal(t, "f-direct", out[0].Fragment.FragmentID)
-	require.Equal(t, 1, expander.calls)
-	require.Equal(t, "pA", expander.lastProfile)
-}
-
-func TestRecallService_CommunityExpansionFiltersProfile(t *testing.T) {
-	now := time.Date(2026, 6, 15, 12, 0, 0, 0, time.UTC)
-	expander := &fakeCommunityExpander{
-		expansion: CommunityExpansion{
-			SelectedCommunities: 1,
-			Facts: []FactRecallResult{
-				{FactID: "fact-other", ProfileID: "pB", RecordedAt: now},
-				{FactID: "fact-own", ProfileID: "pA", RecordedAt: now},
-			},
-			Fragments: []CommunityFragmentRecallResult{
-				{FragmentID: "frag-other", ProfileID: "pB", Score: 0.5},
-				{FragmentID: "frag-own", ProfileID: "pA", Score: 0.4},
-			},
-		},
-	}
-	svc := NewRecallServiceWithTiers(
-		&stubEmbedding{DimensionsResult: 4},
-		&fakeSemanticSearcher{},
-		&fakeKeywordSearcher{},
-		&fakeHydrator{frags: map[string]*domain.Fragment{
-			"frag-other": {FragmentID: "frag-other", ProfileID: "pB"},
-			"frag-own":   {FragmentID: "frag-own", ProfileID: "pA"},
-		}},
-		nil,
-		&fakeFactGetter{facts: map[string]*domain.Fact{
-			"fact-other": {FactID: "fact-other", ProfileID: "pB", Status: domain.FactStatusActive, RecordedAt: now, TruthScore: 1},
-			"fact-own":   {FactID: "fact-own", ProfileID: "pA", Status: domain.FactStatusActive, RecordedAt: now, TruthScore: 0.9},
-		}},
-		nil,
-		nil,
-		0,
-		nil,
-		nil,
-		WithCommunityExpander(expander),
-	)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: "q", Limit: 5, UseCommunities: true})
-	require.NoError(t, err)
-	require.Len(t, out, 2)
-	require.NotNil(t, out[0].Fact)
-	require.Equal(t, "fact-own", out[0].Fact.FactID)
-	require.NotNil(t, out[1].Fragment)
-	require.Equal(t, "frag-own", out[1].Fragment.FragmentID)
-}
-
-func TestRecallService_CommunityExpansionFillsUnusedSlotsAfterDirectRecall(t *testing.T) {
-	now := time.Date(2026, 6, 15, 12, 0, 0, 0, time.UTC)
-	sem := &fakeSemanticSearcher{hits: []semanticsearch.SearchHit{{ID: "f-direct", Type: "fragment"}}}
-	expander := &fakeCommunityExpander{
-		expansion: CommunityExpansion{
-			SelectedCommunities: 1,
-			Facts: []FactRecallResult{
-				{FactID: "fact-community", ProfileID: "pA", RecordedAt: now},
-			},
-			Claims: []ClaimRecallResult{
-				{ClaimID: "claim-community", ProfileID: "pA", RecordedAt: now},
-			},
-			Fragments: []CommunityFragmentRecallResult{
-				{FragmentID: "fragment-community", ProfileID: "pA", Score: 0.7},
-			},
-		},
-	}
-	svc := NewRecallServiceWithTiers(
-		&stubEmbedding{DimensionsResult: 4},
-		sem,
-		&fakeKeywordSearcher{},
-		&fakeHydrator{frags: map[string]*domain.Fragment{
-			"f-direct":           {FragmentID: "f-direct", ProfileID: "pA"},
-			"fragment-community": {FragmentID: "fragment-community", ProfileID: "pA"},
-		}},
-		nil,
-		&fakeFactGetter{facts: map[string]*domain.Fact{
-			"fact-community": {FactID: "fact-community", ProfileID: "pA", Status: domain.FactStatusActive, RecordedAt: now, TruthScore: 0.9},
-		}},
-		nil,
-		&fakeClaimGetter{claims: map[string]*domain.Claim{
-			"claim-community": {ClaimID: "claim-community", ProfileID: "pA", Status: domain.StatusValidated, RecordedAt: now, ExtractConf: 0.8},
-		}},
-		0,
-		nil,
-		nil,
-		WithCommunityExpander(expander),
-	)
-
-	out, err := svc.Recall(context.Background(), "pA", RecallRequest{Query: "q", Limit: 3, UseCommunities: true})
-	require.NoError(t, err)
-	require.Len(t, out, 3)
-	require.NotNil(t, out[0].Fact)
-	require.Equal(t, "fact-community", out[0].Fact.FactID)
-	require.NotNil(t, out[1].Claim)
-	require.Equal(t, "claim-community", out[1].Claim.ClaimID)
-	require.NotNil(t, out[2].Fragment)
-	require.Equal(t, "f-direct", out[2].Fragment.FragmentID)
-	require.Equal(t, DefaultCommunityExpansionCommunityLimit, expander.lastOptions.CommunityLimit)
-	require.Equal(t, DefaultCommunityExpansionMembersPerCommunity, expander.lastOptions.MembersPerCommunity)
-	require.Equal(t, DefaultCommunityExpansionCommunityLimit*DefaultCommunityExpansionMembersPerCommunity, expander.lastOptions.MaxCandidates)
 }
 
 func idsOf(hits []RecallHit) []string {
