--- tests/eval/seeds/local_tiered_v1/cases.jsonl
@@ -1,17 +1,0 @@
-{"case_id":"local_tiered_v1_fact_currentness","query":"What is the current owner for service TIER-F-001?","task_type":"recall","difficulty":"tiered","slices":["tiered","fact_currentness","adversarial"],"expected_behavior":"Rank the newer active fact for service TIER-F-001 first even when the older active fact has a higher truth score.","limit":1}
-{"case_id":"local_tiered_v1_fact_content_date_currentness","query":"Who is the current owner for service TIER-D-001?","task_type":"recall","difficulty":"tiered","slices":["tiered","fact_currentness","typed_content_date","temporal","adversarial"],"expected_behavior":"Rank the fact whose triple text contains the newer month-name date first when valid_from is absent, even if the older fact was imported later and has a higher score.","limit":1}
-{"case_id":"local_tiered_v1_fact_evidence_date_currentness","query":"Who is the current owner for service TIER-S-001?","task_type":"recall","difficulty":"tiered","slices":["tiered","fact_currentness","typed_evidence_date","temporal","adversarial"],"expected_behavior":"Rank the fact whose supporting fragment contains the newer month-name date first when valid_from and triple date text are absent, even if the older fact was imported later and has a higher score.","limit":1}
-{"case_id":"local_tiered_v1_claim_currentness","query":"What is the latest pager for TIER-C-001?","task_type":"recall","difficulty":"tiered","slices":["tiered","claim_currentness","adversarial"],"expected_behavior":"Rank the newer validated claim for TIER-C-001 first even when the older claim has a higher extraction confidence.","limit":1}
-{"case_id":"local_tiered_v1_claim_content_date_currentness","query":"What is the latest pager for service TIER-D-002?","task_type":"recall","difficulty":"tiered","slices":["tiered","claim_currentness","typed_content_date","temporal","adversarial"],"expected_behavior":"Rank the claim whose triple text contains the newer month-name date first when valid_from is absent, even if the older claim was imported later and has a higher score.","limit":1}
-{"case_id":"local_tiered_v1_claim_evidence_date_currentness","query":"What is the latest pager for service TIER-S-002?","task_type":"recall","difficulty":"tiered","slices":["tiered","claim_currentness","typed_evidence_date","temporal","adversarial"],"expected_behavior":"Rank the claim whose supporting fragment contains the newer month-name date first when valid_from and triple date text are absent, even if the older claim was imported later and has a higher score.","limit":1}
-{"case_id":"local_tiered_v1_fact_over_fragment","query":"What is the active escalation owner for service TIER-M-001?","task_type":"recall","difficulty":"tiered","slices":["tiered","fact_over_fragment","cross_tier"],"expected_behavior":"Prefer the active promoted fact over a stale fragment for the same service.","limit":1}
-{"case_id":"local_tiered_v1_claim_over_fragment","query":"What is the latest pager for service TIER-M-002?","task_type":"recall","difficulty":"tiered","slices":["tiered","claim_over_fragment","cross_tier"],"expected_behavior":"Prefer the validated claim over a stale fragment when no active fact exists for the service.","limit":1}
-{"case_id":"local_tiered_v1_cross_identifier_claim","query":"What is the latest pager for service TIER-X-001?","task_type":"recall","difficulty":"tiered","slices":["tiered","cross_identifier","adversarial"],"expected_behavior":"Return the claim for TIER-X-001 and do not let a neighboring TIER-X-999 fact win by tier precedence.","limit":1}
-{"case_id":"local_tiered_v1_fact_valid_window","query":"As of 2026-06-22, what owner did service TIER-W-001 use?","task_type":"recall","difficulty":"tiered","slices":["tiered","fact_valid_window","temporal"],"expected_behavior":"Return the fact valid on 2026-06-22 and ignore a stronger future fact from 2026-06-27.","valid_at":"2026-06-22T00:00:00Z","limit":1}
-{"case_id":"local_tiered_v1_claim_valid_window","query":"As of 2026-06-22, what pager did service TIER-W-002 use?","task_type":"recall","difficulty":"tiered","slices":["tiered","claim_valid_window","temporal"],"expected_behavior":"Return the claim valid on 2026-06-22 and ignore a stronger future claim from 2026-06-27.","valid_at":"2026-06-22T00:00:00Z","limit":1}
-{"case_id":"local_tiered_v1_evidence_source_fragment","query":"Which source note says service TIER-E-001 owner uses owner-lumen?","task_type":"recall","difficulty":"tiered","slices":["tiered","evidence_source","cross_tier"],"expected_behavior":"Return the raw source fragment when the query asks which source note says the fact, not the derived fact or claim.","limit":1}
-{"case_id":"local_tiered_v1_fragment_relative_currentness","query":"Who is the current owner for service TIER-R-001?","task_type":"recall","difficulty":"tiered","slices":["tiered","fragment_currentness","temporal","adversarial"],"expected_behavior":"Prefer the fragment with a relative dated update over an undated fragment that only says current.","limit":1}
-{"case_id":"local_tiered_v1_fragment_month_name_currentness","query":"Who is the current owner for service TIER-N-001?","task_type":"recall","difficulty":"tiered","slices":["tiered","fragment_currentness","temporal","adversarial"],"expected_behavior":"Prefer the fragment with a month-name dated update over an undated fragment that only says current.","limit":1}
-{"case_id":"local_tiered_v1_fragment_weekday_currentness","query":"Who is the current owner for service TIER-Y-001?","task_type":"recall","difficulty":"tiered","slices":["tiered","fragment_currentness","temporal","adversarial"],"expected_behavior":"Prefer the fragment with a weekday dated update over an undated fragment that only says current.","limit":1}
-{"case_id":"local_tiered_v1_fragment_numeric_currentness","query":"Who is the current owner for service TIER-Q-001?","task_type":"recall","difficulty":"tiered","slices":["tiered","fragment_currentness","temporal","adversarial"],"expected_behavior":"Prefer the fragment with an unambiguous numeric dated update over an undated fragment that only says current.","limit":1}
-{"case_id":"local_tiered_v1_fragment_expired_range_v2_currentness","query":"Who is the current owner for service TIER-RG-777?","task_type":"recall","difficulty":"tiered","slices":["tiered","fragment_currentness","temporal_range","adversarial"],"expected_behavior":"Prefer the current fragment over a same-identifier fragment whose only date is an expired valid-through range.","limit":1}
