--- tests/eval/seeds/local_tiered_v1/licenses.md
@@ -1,5 +1,0 @@
-# local_tiered_v1 Licenses
-
-This seed is synthetic project test data authored for Dense-Mem recall evaluation.
-
-License: Apache-2.0
