--- tests/eval/seeds/local_tiered_v1/qrels.jsonl
@@ -1,17 +1,0 @@
-{"case_id":"local_tiered_v1_fact_currentness","required_refs":[{"type":"fact","source_doc_id":"local_tiered_v1_fact_new","grade":3,"reason":"newer current active fact"}],"bad_refs":[{"type":"fact","source_doc_id":"local_tiered_v1_fact_old","reason":"older same-tier fact should not outrank the current fact"}]}
-{"case_id":"local_tiered_v1_fact_content_date_currentness","required_refs":[{"type":"fact","source_doc_id":"local_tiered_v1_fact_content_date_required","grade":3,"reason":"fact triple text carries the newer month-name date"}],"bad_refs":[{"type":"fact","source_doc_id":"local_tiered_v1_fact_content_date_negative","reason":"older content-dated fact should not outrank the newer content-dated fact even when imported later"}]}
-{"case_id":"local_tiered_v1_fact_evidence_date_currentness","required_refs":[{"type":"fact","source_doc_id":"local_tiered_v1_fact_evidence_date_required","grade":3,"reason":"supporting fragment text carries the newer month-name date for this fact"}],"bad_refs":[{"type":"fact","source_doc_id":"local_tiered_v1_fact_evidence_date_negative","reason":"older evidence-dated fact should not outrank the newer evidence-dated fact even when imported later"}],"required_evidence_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fact_evidence_date_required","grade":3,"reason":"returned fact must be grounded by the newer source fragment"}],"bad_evidence_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fact_evidence_date_negative","reason":"older source fragment must not ground the returned fact context"}]}
-{"case_id":"local_tiered_v1_claim_currentness","required_refs":[{"type":"claim","source_doc_id":"local_tiered_v1_claim_new","grade":3,"reason":"newer current validated claim"}],"bad_refs":[{"type":"claim","source_doc_id":"local_tiered_v1_claim_old","reason":"older same-tier claim should not outrank the current claim"}]}
-{"case_id":"local_tiered_v1_claim_content_date_currentness","required_refs":[{"type":"claim","source_doc_id":"local_tiered_v1_claim_content_date_required","grade":3,"reason":"claim triple text carries the newer month-name date"}],"bad_refs":[{"type":"claim","source_doc_id":"local_tiered_v1_claim_content_date_negative","reason":"older content-dated claim should not outrank the newer content-dated claim even when imported later"}]}
-{"case_id":"local_tiered_v1_claim_evidence_date_currentness","required_refs":[{"type":"claim","source_doc_id":"local_tiered_v1_claim_evidence_date_required","grade":3,"reason":"supporting fragment text carries the newer month-name date for this claim"}],"bad_refs":[{"type":"claim","source_doc_id":"local_tiered_v1_claim_evidence_date_negative","reason":"older evidence-dated claim should not outrank the newer evidence-dated claim even when imported later"}],"required_evidence_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_claim_evidence_date_required","grade":3,"reason":"returned claim must be grounded by the newer source fragment"}],"bad_evidence_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_claim_evidence_date_negative","reason":"older source fragment must not ground the returned claim context"}]}
-{"case_id":"local_tiered_v1_fact_over_fragment","required_refs":[{"type":"fact","source_doc_id":"local_tiered_v1_fact_over_fragment_required","grade":3,"reason":"active promoted fact should answer the service owner query"}],"bad_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fact_over_fragment_negative","reason":"stale fragment should not outrank the active fact"}]}
-{"case_id":"local_tiered_v1_claim_over_fragment","required_refs":[{"type":"claim","source_doc_id":"local_tiered_v1_claim_over_fragment_required","grade":3,"reason":"validated claim should answer the pager query"}],"bad_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_claim_over_fragment_negative","reason":"stale fragment should not outrank the validated claim"}]}
-{"case_id":"local_tiered_v1_cross_identifier_claim","required_refs":[{"type":"claim","source_doc_id":"local_tiered_v1_cross_identifier_claim_required","grade":3,"reason":"validated claim matches the requested TIER-X-001 identifier"}],"bad_refs":[{"type":"fact","source_doc_id":"local_tiered_v1_cross_identifier_fact_negative","reason":"neighboring TIER-X-999 fact should be rejected by identifier matching"}]}
-{"case_id":"local_tiered_v1_fact_valid_window","required_refs":[{"type":"fact","source_doc_id":"local_tiered_v1_fact_valid_window_old","grade":3,"reason":"fact is valid at the requested 2026-06-22 time"}],"bad_refs":[{"type":"fact","source_doc_id":"local_tiered_v1_fact_valid_window_future","reason":"future 2026-06-27 fact must not answer a 2026-06-22 query"}]}
-{"case_id":"local_tiered_v1_claim_valid_window","required_refs":[{"type":"claim","source_doc_id":"local_tiered_v1_claim_valid_window_old","grade":3,"reason":"claim is valid at the requested 2026-06-22 time"}],"bad_refs":[{"type":"claim","source_doc_id":"local_tiered_v1_claim_valid_window_future","reason":"future 2026-06-27 claim must not answer a 2026-06-22 query"}]}
-{"case_id":"local_tiered_v1_evidence_source_fragment","required_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_evidence_source_required","grade":3,"reason":"source-intent query asks for the original note"}],"bad_refs":[{"type":"fact","source_doc_id":"local_tiered_v1_evidence_source_required","reason":"derived fact answers the value but not which source note said it"},{"type":"claim","source_doc_id":"local_tiered_v1_evidence_source_required","reason":"derived claim answers the value but not which source note said it"},{"type":"fragment","source_doc_id":"local_tiered_v1_evidence_source_negative","reason":"neighbor source note is for TIER-E-999, not TIER-E-001"}]}
-{"case_id":"local_tiered_v1_fragment_relative_currentness","required_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fragment_relative_required","grade":3,"reason":"relative dated owner update is newer evidence than the undated current note"}],"bad_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fragment_relative_negative","reason":"undated current wording should not outrank the relative dated update"}]}
-{"case_id":"local_tiered_v1_fragment_month_name_currentness","required_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fragment_month_name_required","grade":3,"reason":"month-name dated owner update is newer evidence than the undated current note"}],"bad_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fragment_month_name_negative","reason":"undated current wording should not outrank the month-name dated update"}]}
-{"case_id":"local_tiered_v1_fragment_weekday_currentness","required_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fragment_weekday_required","grade":3,"reason":"weekday dated owner update is newer evidence than the undated current note"}],"bad_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fragment_weekday_negative","reason":"undated current wording should not outrank the weekday dated update"}]}
-{"case_id":"local_tiered_v1_fragment_numeric_currentness","required_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fragment_numeric_required","grade":3,"reason":"numeric dated owner update is newer evidence than the undated current note"}],"bad_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fragment_numeric_negative","reason":"undated current wording should not outrank the numeric dated update"}]}
-{"case_id":"local_tiered_v1_fragment_expired_range_v2_currentness","required_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fragment_expired_range_v2_required","grade":3,"reason":"current same-identifier owner note remains active"}],"bad_refs":[{"type":"fragment","source_doc_id":"local_tiered_v1_fragment_expired_range_v2_negative","reason":"valid-through range ended before the current query and must not be treated as a fresh update date"}]}
