--- tests/eval/seeds/local_tiered_v1/seed_manifest.json
@@ -1,25 +1,0 @@
-{
-  "schema_version": "dense-mem.eval.seed.v1",
-  "seed_id": "local_tiered_v1",
-  "description": "Small local eval seed for fact and claim recall tiers, focused on same-tier currentness ordering, typed content-date currentness, typed evidence-date currentness, cross-tier fact/claim/fragment behavior, valid_at temporal windows, evidence-source intent, and natural-language temporal fragment currentness including relative, month-name, weekday, numeric, and expired range cues.",
-  "generated_at": "2026-06-28T22:50:00Z",
-  "corpus_file": "corpus.jsonl",
-  "cases_file": "cases.jsonl",
-  "qrels_file": "qrels.jsonl",
-  "licenses_file": "licenses.md",
-  "embedding_provider": "configured-runtime",
-  "embedding_model": "configured-runtime",
-  "counts": {
-    "corpus": 34,
-    "cases": 17,
-    "qrels": 17,
-    "docs_per_case": 2
-  },
-  "sources": [
-    {
-      "name": "Dense-Mem synthetic tiered recall seed",
-      "license": "Apache-2.0",
-      "notes": "Synthetic fixture authored for local recall evaluation."
-    }
-  ]
-}
