--- tests/eval/suites/local_tiered_v1.jsonl
@@ -1,17 +1,0 @@
-{"case_id":"local_tiered_v1_fact_currentness","weight":1,"slices":["tiered","fact_currentness","adversarial"]}
-{"case_id":"local_tiered_v1_fact_content_date_currentness","weight":1,"slices":["tiered","fact_currentness","typed_content_date","temporal","adversarial"]}
-{"case_id":"local_tiered_v1_fact_evidence_date_currentness","weight":1,"slices":["tiered","fact_currentness","typed_evidence_date","temporal","adversarial"]}
-{"case_id":"local_tiered_v1_claim_currentness","weight":1,"slices":["tiered","claim_currentness","adversarial"]}
-{"case_id":"local_tiered_v1_claim_content_date_currentness","weight":1,"slices":["tiered","claim_currentness","typed_content_date","temporal","adversarial"]}
-{"case_id":"local_tiered_v1_claim_evidence_date_currentness","weight":1,"slices":["tiered","claim_currentness","typed_evidence_date","temporal","adversarial"]}
-{"case_id":"local_tiered_v1_fact_over_fragment","weight":1,"slices":["tiered","fact_over_fragment","cross_tier"]}
-{"case_id":"local_tiered_v1_claim_over_fragment","weight":1,"slices":["tiered","claim_over_fragment","cross_tier"]}
-{"case_id":"local_tiered_v1_cross_identifier_claim","weight":1,"slices":["tiered","cross_identifier","adversarial"]}
-{"case_id":"local_tiered_v1_fact_valid_window","weight":1,"slices":["tiered","fact_valid_window","temporal"]}
-{"case_id":"local_tiered_v1_claim_valid_window","weight":1,"slices":["tiered","claim_valid_window","temporal"]}
-{"case_id":"local_tiered_v1_evidence_source_fragment","weight":1,"slices":["tiered","evidence_source","cross_tier"]}
-{"case_id":"local_tiered_v1_fragment_relative_currentness","weight":1,"slices":["tiered","fragment_currentness","temporal","adversarial"]}
-{"case_id":"local_tiered_v1_fragment_month_name_currentness","weight":1,"slices":["tiered","fragment_currentness","temporal","adversarial"]}
-{"case_id":"local_tiered_v1_fragment_weekday_currentness","weight":1,"slices":["tiered","fragment_currentness","temporal","adversarial"]}
-{"case_id":"local_tiered_v1_fragment_numeric_currentness","weight":1,"slices":["tiered","fragment_currentness","temporal","adversarial"]}
-{"case_id":"local_tiered_v1_fragment_expired_range_v2_currentness","weight":1,"slices":["tiered","fragment_currentness","temporal_range","adversarial"]}
